refactor(timeline): split the panes and remaining DSL out of room_screen mod.rs
Stage 3 of the room_screen split, and the first stage that moves DSL:
the threads pane (threads_pane.rs), room info pane (room_info_pane.rs),
report-room modal (report_modal.rs), and app-service panel (into
bot_admin.rs) each take their widgets AND their script_mod! templates;
the rest of the DSL block (colors, message template family, small-state
rows, Timeline/RoomScreen screen templates) moves to dsl.rs.

mod.rs now registers the blocks through a hand-written script_mod in
dependency order — panes first, since the RoomScreen template in `dsl`
references them via mod.widgets.* and a template must already be
registered when a later block resolves such a reference. Verified by
booting the app: all templates register with no resolution errors.

Pure code motion otherwise; the public surface is preserved via pub use
re-exports. mod.rs: 12,659 -> 7,903 lines.
