+++ src/home/room_screen/mod.rs
@@ -60,15 +60,22 @@ use super::{ContextMenuOpenGesture, event_reaction_list::ReactionData, invite_mo
 
 mod bot_admin;
 mod bot_message;
+mod dsl;
 mod octos_actions;
+mod report_modal;
+mod room_info_pane;
 mod small_state;
 mod state;
+mod threads_pane;
 
-use bot_admin::*;
+pub use bot_admin::*;
 use bot_message::*;
 use octos_actions::*;
+pub use report_modal::*;
+pub use room_info_pane::*;
 use small_state::*;
 use state::*;
+pub use threads_pane::*;
 pub(crate) use bot_admin::is_known_or_likely_bot;
 pub use state::{RoomScreenProps, RoomScreenTooltipActions, TimelineUpdate, clear_timeline_states};
 
@@ -266,21 +273,6 @@ fn wedge_error_reason(error: &matrix_sdk::Error) -> String {
     }
 }
 
-thread_local! {
-    static ROOM_INFO_ACTION_MODAL_OPEN: Cell<bool> = const { Cell::new(false) };
-}
-
-/// Set by app.rs each frame from the GLOBAL room-info-action modals
-/// (report / leave-confirm), so the room info sliding pane knows not to
-/// self-close on Escape / tap-outside while one of them is open over it.
-pub fn set_room_info_action_modal_open(open: bool) {
-    ROOM_INFO_ACTION_MODAL_OPEN.with(|state| state.set(open));
-}
-
-fn is_room_info_action_modal_open() -> bool {
-    ROOM_INFO_ACTION_MODAL_OPEN.with(|state| state.get())
-}
-
 
 /// #FFF4E5
 const COLOR_THREAD_SUMMARY_BG: Vec4 = vec4(1.0, 0.957, 0.898, 1.0);
@@ -317,4371 +309,17 @@ fn compute_translation_lang_popup_abs_pos(button_rect: Rect, container_rect: Rec
     dvec2(popup_x, popup_y)
 }
 
-script_mod! {
-    use mod.prelude.widgets.*
-    use mod.widgets.*
-
-
-    mod.widgets.COLOR_BG = #xfff8ee
-    mod.widgets.COLOR_OVERLAY_BG = #x000000d8
-    mod.widgets.COLOR_READ_MARKER = #xeb2733
-
-    mod.widgets.REACTION_TEXT_COLOR = #4c00b0
-
-    mod.widgets.COLOR_THREAD_SUMMARY_BG = #FFF4E5
-    mod.widgets.COLOR_THREAD_SUMMARY_BG_HOVER = #FFEACC
-    mod.widgets.COLOR_THREAD_SUMMARY_BORDER = #E8C99A
-    mod.widgets.COLOR_THREAD_SUMMARY_REPLY_COUNT = #A35A00
-    // Aliases onto the token layer. These were seven hand-picked blue-greys that
-    // belonged to no ramp — each within ~2% of a real token, which is exactly
-    // the drift that makes an agent card read as slightly "off" next to the
-    // approval card and every other surface nested inside it.
-    mod.widgets.COLOR_BOT_CARD_BG = (mod.widgets.RBX_BG_CANVAS)
-    mod.widgets.COLOR_BOT_CARD_BORDER = (mod.widgets.RBX_STROKE_STRONG)
-    mod.widgets.COLOR_BOT_STATUS_BG = (mod.widgets.RBX_BG_SUNKEN)
-    mod.widgets.COLOR_BOT_STATUS_TEXT = (mod.widgets.RBX_FG_SECONDARY)
-    mod.widgets.COLOR_BOT_FOOTER_TEXT = (mod.widgets.RBX_FG_TERTIARY)
-    mod.widgets.COLOR_BOT_CODE_BG = (mod.widgets.RBX_BG_SUNKEN)
-    mod.widgets.COLOR_BOT_CODE_BORDER = (mod.widgets.RBX_STROKE_STRONG)
-
-    mod.widgets.MessageActionPrimaryButton = RobrixPositiveIconButton {
-        width: Fit
-        height: Fit
-        spacing: 6.0
-        padding: Inset{ left: 10.0, right: 10.0, top: 7.0, bottom: 7.0 }
-        draw_text +: {
-            text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 10.0 }
-        }
-    }
-
-    mod.widgets.MessageActionSecondaryButton = Button {
-        width: Fit
-        height: Fit
-        spacing: 6.0
-        padding: Inset{ left: 10.0, right: 10.0, top: 7.0, bottom: 7.0 }
-        draw_text +: {
-            text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 10.0 }
-        }
-        text: ""
-    }
-
-    mod.widgets.MessageActionDangerButton = RobrixNegativeIconButton {
-        width: Fit
-        height: Fit
-        spacing: 6.0
-        padding: Inset{ left: 10.0, right: 10.0, top: 7.0, bottom: 7.0 }
-        draw_text +: {
-            text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 10.0 }
-        }
-    }
-
-    mod.widgets.SmallStateGroupToggleButton = Button {
-        width: Fit
-        height: Fit
-        spacing: 0.0
-        padding: Inset{ left: 0.0, right: 0.0, top: 0.0, bottom: 0.0 }
-        draw_bg +: {
-            border_radius: 4.0
-            border_size: 0.0
-            border_color: #x00000000
-            border_color_hover: #x00000000
-            border_color_down: #x00000000
-            color: #x00000000
-            color_hover: #x00000000
-            color_down: #x00000000
-        }
-        draw_text +: {
-            text_style: SMALL_STATE_TEXT_STYLE { font_size: 11.0 }
-            color: #x232A31
-            color_hover: #x1A1F25
-            color_down: #x0E1217
-        }
-        text: ""
-    }
-
-    mod.widgets.MessageActionButtonSlot = View {
-        visible: false
-        width: Fit
-        height: Fit
-        flow: Overlay
-
-        primary_button := mod.widgets.MessageActionPrimaryButton {
-            visible: false
-        }
-        secondary_button := mod.widgets.MessageActionSecondaryButton {
-            visible: false
-        }
-        danger_button := mod.widgets.MessageActionDangerButton {
-            visible: false
-        }
-    }
-
-    mod.widgets.BotTimelineMarkdown = Markdown {
-        width: Fill
-        height: Fit
-        padding: 0.0
-        font_size: (MESSAGE_FONT_SIZE)
-        font_color: (MESSAGE_TEXT_COLOR)
-        paragraph_spacing: 10.0
-        pre_code_spacing: 8.0
-        heading_base_scale: 1.45
-        inline_code_padding: Inset{ top: 3, bottom: 3, left: 4, right: 4 }
-        inline_code_margin: Inset{ left: 3, right: 3, bottom: 2, top: 2 }
-        use_code_block_widget: true
-
-        draw_text +: {
-            color: (MESSAGE_TEXT_COLOR)
-        }
-        text_style_normal: mod.widgets.MESSAGE_TEXT_STYLE {
-            font_size: (MESSAGE_FONT_SIZE)
-            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-        }
-        text_style_italic: theme.font_italic {
-            font_size: (MESSAGE_FONT_SIZE)
-            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-        }
-        text_style_bold: theme.font_bold {
-            font_size: (MESSAGE_FONT_SIZE)
-            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-        }
-        text_style_bold_italic: theme.font_bold_italic {
-            font_size: (MESSAGE_FONT_SIZE)
-            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-        }
-        text_style_fixed: mod.widgets.MESSAGE_CODE_TEXT_STYLE {
-            font_size: (MESSAGE_FONT_SIZE - 0.5)
-            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-        }
-        draw_block +: {
-            line_color: (MESSAGE_TEXT_COLOR)
-            sep_color: (mod.widgets.COLOR_BOT_CODE_BORDER)
-            quote_bg_color: (mod.widgets.RBX_BG_SUNKEN)
-            quote_fg_color: (mod.widgets.RBX_FG_SECONDARY)
-            code_color: (mod.widgets.COLOR_BOT_CODE_BG)
-        }
-        code_layout: Layout{
-            flow: Flow.Right{wrap: true}
-            padding: Inset{ left: 0.0, right: 0.0, top: 0.0, bottom: 0.0 }
-        }
-        code_walk: Walk{ width: Fill, height: Fit, margin: Inset{ top: 10.0, bottom: 10.0 } }
-        quote_layout: Layout{
-            flow: Flow.Right{wrap: true}
-            padding: Inset{ left: 12.0, right: 12.0, top: 8.0, bottom: 8.0 }
-        }
-        quote_walk: Walk{ width: Fill, height: Fit, margin: Inset{ top: 6.0, bottom: 6.0 } }
-        list_item_layout: Layout{
-            flow: Flow.Right{wrap: true}
-            padding: Inset{ left: 0.0, right: 0.0, top: 1.0, bottom: 1.0 }
-        }
-        list_item_walk: Walk{ width: Fill, height: Fit, margin: Inset{ top: 0.0, bottom: 1.0 } }
-
-        code_block := RoundedView {
-            width: Fill
-            height: Fit
-            flow: Overlay
-            padding: 0.0
-            show_bg: true
-            draw_bg +: {
-                color: (mod.widgets.COLOR_BOT_CODE_BG)
-                border_radius: 10.0
-                border_size: 1.0
-                border_color: (mod.widgets.COLOR_BOT_CODE_BORDER)
-            }
-
-            code_view := mod.widgets.CodeView {
-                keep_cursor_at_end: false
-                editor +: {
-                    width: Fill
-                    height: Fit
-                    margin: Inset{ left: 12.0, right: 12.0, top: 10.0, bottom: 10.0 }
-                    draw_bg +: { color: #0000 }
-                    draw_text +: {
-                        text_style: mod.widgets.MESSAGE_CODE_TEXT_STYLE {
-                            font_size: (MESSAGE_FONT_SIZE - 0.5)
-                            line_spacing: (MESSAGE_TEXT_LINE_SPACING)
-                        }
-                    }
-                    token_colors +: {
-                        whitespace: #x6a737d
-                        delimiter: #x24292e
-                        delimiter_highlight: #x005cc5
-                        error_decoration: #xcb2431
-                        warning_decoration: #xb08800
-                        unknown: #x24292e
-                        branch_keyword: #xd73a49
-                        constant: #x005cc5
-                        identifier: #x24292e
-                        loop_keyword: #xd73a49
-                        number: #x005cc5
-                        other_keyword: #xd73a49
-                        punctuator: #x24292e
-                        string: #x22863a
-                        function: #x6f42c1
-                        typename: #xe36209
-                        comment: #x6a737d
-                    }
-                }
-            }
-        }
-    }
-
-    // An empty view that takes up no space in the portal list.
-    mod.widgets.Empty = View { }
-
-    mod.widgets.MessageDownloadSection = View {
-        visible: false,
-        width: Fit, height: Fit,
-        flow: Right,
-        margin: Inset{top: 8, bottom: 2}
-
-        download_button := RobrixIconButton {
-            height: mod.widgets.SETTINGS_BUTTON_HEIGHT,
-            padding: Inset{left: 12, right: 12}
-            margin: 0
-            draw_icon.svg: (ICON_DOWNLOAD)
-            icon_walk: Walk{width: 16, height: 16}
-            text: "Download"
-        }
-
-        downloading_view := View {
-            visible: false,
-            width: Fit, height: mod.widgets.SETTINGS_BUTTON_HEIGHT
-            flow: Right,
-            align: Align{y: 0.5}
-            spacing: 8,
-            padding: Inset{left: 12, right: 6}
-
-            spinner := LoadingSpinner {
-                width: 16, height: 16
-                draw_bg.color: (COLOR_ACTIVE_PRIMARY)
-            }
-            status_label := Label {
-                width: Fit, height: Fit
-                padding: 0
-                margin: 0
-                draw_text +: {
-                    text_style: REGULAR_TEXT { font_size: 11 },
-                    color: (COLOR_ACTIVE_PRIMARY)
-                }
-                text: "Downloading…"
-            }
-            cancel_button := RobrixNegativeIconButton {
-                height: mod.widgets.SETTINGS_BUTTON_HEIGHT,
-                padding: Inset{left: 12, right: 12}
-                margin: 0
-                draw_icon.svg: (ICON_CLOSE)
-                icon_walk: Walk{width: 16, height: 16}
-                text: "Cancel"
-            }
-        }
-
-        success_button := RobrixPositiveIconButton {
-            visible: false,
-            height: mod.widgets.SETTINGS_BUTTON_HEIGHT,
-            padding: Inset{left: 12, right: 12}
-            margin: 0
-            draw_icon.svg: (ICON_CHECKMARK)
-            icon_walk: Walk{width: 16, height: 16}
-            text: "Downloaded"
-        }
-
-        failure_button := RobrixNegativeIconButton {
-            visible: false,
-            height: mod.widgets.SETTINGS_BUTTON_HEIGHT,
-            padding: Inset{left: 12, right: 12}
-            margin: 0
-            draw_icon.svg: (ICON_CLOSE)
-            icon_walk: Walk{width: 16, height: 16}
-            text: "Download Failed"
-        }
-    }
-
-    // A summary at the bottom of a message that is the root of a thread.
-    mod.widgets.ThreadRootSummary = RoundedView {
-        visible: false
-        width: Fill,
-        height: Fit
-        flow: Right,
-        align: Align{x: 0.0, y: 0.5}
-        spacing: 5.0
-        margin: Inset{ top: 5.0 }
-        padding: 12,
-        cursor: MouseCursor.Hand
-
-        show_bg: true
-        draw_bg +: {
-            color: (mod.widgets.COLOR_THREAD_SUMMARY_BG)
-            border_radius: 4.0
-            border_size: 1.5
-            border_color: (mod.widgets.COLOR_THREAD_SUMMARY_BORDER)
-        }
-
-        thread_summary_count := Label {
-            width: Fit,
-            draw_text +: {
-                text_style: USERNAME_TEXT_STYLE { font_size: 11 }
-                color: (mod.widgets.COLOR_THREAD_SUMMARY_REPLY_COUNT)
-            }
-            text: ""
-        }
-
-        Icon {
-            width: Fit, height: Fit,
-            align: Align{x: 0.5, y: 0.5}
-            draw_icon +: {
-                svg: crate_resource("self://resources/icons/double_chat.svg")
-                color: (mod.widgets.COLOR_THREAD_SUMMARY_REPLY_COUNT)
-            }
-            icon_walk: Walk{ width: 25, height: 25, margin: Inset{top: 3, right: 7} }
-        }
-
-        thread_summary_latest := MessageHtml {
-            flow: Right,
-            max_lines: 2
-            text_overflow: Ellipsis
-        }
-    }
-
-    // The view used for each text-based message event in a room's timeline.
-    // The per-message meta band: copy action (left) · bot model metadata
-    // (middle) · read receipts (right). Declared once and instantiated by both
-    // the Message and CondensedMessage templates — CondensedMessage re-declares
-    // its whole `body` subtree with `:=`, so it cannot inherit this from Message.
-    // (Named MessageMetaBand to avoid colliding with the retired floating
-    // MessageActionBar popup referenced in a commented-out block below.)
-    // Shown under a message whose send is parked and will not move again on its
-    // own: why it failed, and the two ways out. Declared once and instantiated
-    // by both message templates, like MessageDownloadSection.
-    mod.widgets.SendFailureSection = View {
-        visible: false
-        width: Fill,
-        height: Fit
-        flow: Down,
-        spacing: (SPACE_XS)
-        margin: Inset{ top: 4.0, bottom: 2.0 }
-
-        send_failure_reason := Label {
-            width: Fill,
-            height: Fit
-            flow: Flow.Right{wrap: true}
-            draw_text +: {
-                text_style: RBX_TEXT_META {}
-                color: (RBX_DANGER_FG)
-            }
-            text: ""
-        }
-
-        send_failure_actions := View {
-            width: Fill,
-            height: Fit
-            flow: Right,
-            spacing: (SPACE_XS)
-
-            send_retry_button := RobrixIconButton {
-                width: Fit, height: Fit
-                padding: Inset{left: 10.0, right: 10.0, top: 4.0, bottom: 4.0}
-                draw_bg +: {
-                    color: (RBX_ACCENT_SOFT)
-                    color_hover: (RBX_ACCENT_SOFT)
-                    color_down: (RBX_BG_PRESSED)
-                    border_radius: (RBX_RADIUS_MD)
-                    border_size: 1.0
-                    border_color: (RBX_ACCENT)
-                    border_color_hover: (RBX_ACCENT)
-                    border_color_down: (RBX_ACCENT)
-                }
-                draw_text +: {
-                    text_style: (RBX_TEXT_META)
-                    color: (RBX_ACCENT)
-                    color_hover: (RBX_ACCENT)
-                    color_down: (RBX_ACCENT)
-                    color_focus: (RBX_ACCENT)
-                }
-                draw_icon +: { svg: (ICON_SEND), color: (RBX_ACCENT) }
-                icon_walk: Walk{width: 12, height: 12}
-                text: "Retry"
-            }
-
-            send_discard_button := RobrixIconButton {
-                width: Fit, height: Fit
-                padding: Inset{left: 10.0, right: 10.0, top: 4.0, bottom: 4.0}
-                draw_bg +: {
-                    color: (RBX_DANGER_BG)
-                    color_hover: (RBX_DANGER_BG)
-                    color_down: (RBX_BG_PRESSED)
-                    border_radius: (RBX_RADIUS_MD)
-                    border_size: 1.0
-                    border_color: (RBX_DANGER_FG)
-                    border_color_hover: (RBX_DANGER_FG)
-                    border_color_down: (RBX_DANGER_FG)
-                }
-                draw_text +: {
-                    text_style: (RBX_TEXT_META)
-                    color: (RBX_DANGER_FG)
-                    color_hover: (RBX_DANGER_FG)
-                    color_down: (RBX_DANGER_FG)
-                    color_focus: (RBX_DANGER_FG)
-                }
-                draw_icon +: { svg: (ICON_TRASH), color: (RBX_DANGER_FG) }
-                icon_walk: Walk{width: 12, height: 12}
-                text: "Discard"
-            }
-        }
-    }
-
-    // Delivery-state pill for a message still in the send queue. Shares the
-    // meta band so it costs no vertical space, and sits next to the read
-    // receipts — both answer "did this message land?".
-    mod.widgets.SendStatePill = RoundedView {
-        visible: false
-        width: Fit,
-        height: Fit
-        padding: Inset{ left: 6.0, right: 6.0, top: 1.0, bottom: 1.0 }
-        show_bg: true
-        draw_bg +: {
-            color: (RBX_NEUTRAL_BG)
-            border_radius: (RBX_RADIUS_PILL)
-        }
-
-        send_state_label := Label {
-            width: Fit,
-            height: Fit
-            padding: 0
-            draw_text +: {
-                text_style: RBX_TEXT_META {}
-                color: (RBX_NEUTRAL_FG)
-            }
-            text: ""
-        }
-    }
-
-    mod.widgets.MessageMetaBand = View {
-        width: Fill,
-        height: Fit
-        flow: Right,
-        align: Align{y: 0.5}
-        spacing: (SPACE_XS)
-
-        send_state_pill := mod.widgets.SendStatePill {}
-
-        // Fold affordance for an over-long plain message. Bot replies carry
-        // their own toggle inside the card footer; this is the equivalent for
-        // everything else, and it lives in the meta band so no message template
-        // needs a new slot.
-        plain_fold_toggle := mod.widgets.SmallStateGroupToggleButton {
-            visible: false
-            padding: Inset{ left: 0.0, right: 6.0, top: 0.0, bottom: 0.0 }
-            draw_text +: {
-                text_style: RBX_TEXT_META {}
-                color: (RBX_ACCENT)
-                color_hover: (RBX_ACCENT_HOVER)
-                color_down: (RBX_ACCENT_PRESSED)
-                color_focus: (RBX_ACCENT)
-            }
-            text: ""
-        }
-
-        // Message-type badge (request / reply / info) for bridge-relayed agent
-        // messages, derived from the leading emoji marker the bridge stamps on
-        // the body — the one language-independent signal on the wire. It rides
-        // in the meta band rather than on its own row so it costs no vertical
-        // space. Hidden for every message without that marker.
-        kind_badge := RoundedView {
-            visible: false
-            width: Fit,
-            height: Fit
-            padding: Inset{ left: 6.0, right: 6.0, top: 1.0, bottom: 1.0 }
-            show_bg: true
-            draw_bg +: {
-                color: (RBX_ACCENT_SOFT)
-                border_radius: 3.0
-            }
-
-            kind_badge_label := Label {
-                width: Fit,
-                height: Fit
-                padding: 0
-                draw_text +: {
-                    text_style: RBX_TEXT_META {}
-                    color: (RBX_ACCENT)
-                }
-                text: ""
-            }
-        }
-        copy_button := RobrixNeutralIconButton {
-            visible: false
-            width: Fit,
-            height: Fit,
-            padding: (SPACE_XS)
-            // Optical alignment: cancel the button's own SPACE_XS padding so
-            // the icon glyph lines up with the message text's left edge.
-            margin: Inset{ left: -4 }
-            spacing: 0
-            draw_bg +: {
-                color: (RBX_TRANSPARENT)
-                color_hover: (RBX_HIT_HOVER)
-                color_down: (RBX_HIT_DOWN)
-                border_size: 0.0
-            }
-            draw_icon +: { svg: (ICON_COPY), color: (RBX_FG_TERTIARY) }
-            icon_walk: Walk{width: (RBX_ICON_SM), height: (RBX_ICON_SM)}
-            text: ""
-        }
-        metadata_label := Label {
-            visible: false
-            width: Fill,
-            height: Fit
-            padding: 0
-            max_lines: 1
-            text_overflow: Ellipsis
-            draw_text +: {
-                text_style: RBX_TEXT_META {}
-                color: (RBX_FG_TERTIARY)
-            }
-            text: ""
-        }
-        avatar_row := mod.widgets.AvatarRow {}
-    }
-
-    mod.widgets.Message = set_type_default() do #(Message::register_widget(vm)) {
-
-        width: Fill,
-        height: Fit,
-        margin: 0.0
-        flow: Down,
-        cursor: MouseCursor.Default,
-        padding: 0.0,
-        spacing: 0.0
-
-        show_bg: true
-        draw_bg +: {
-            highlight: instance(0.0)
-            hover: instance(0.0)
-            color: instance((COLOR_PRIMARY)) // default color)
-
-            mentions_bar_color: instance((COLOR_PRIMARY))
-            mentions_bar_width: instance(4.0)
-
-            pixel: fn() {
-                let base_color = mix(
-                    self.color,
-                    #fafafa,
-                    self.hover
-                );
-
-                let with_highlight = mix(
-                    base_color,
-                    #c5d6fa,
-                    self.highlight
-                );
-
-                let sdf = Sdf2d.viewport(self.pos * self.rect_size);
-
-                // draw bg
-                sdf.rect(0., 0., self.rect_size.x, self.rect_size.y);
-                sdf.fill(with_highlight);
-
-                // draw the left vertical line
-                sdf.rect(0., 0., self.mentions_bar_width, self.rect_size.y);
-                sdf.fill(self.mentions_bar_color);
-
-                return sdf.result;
-            }
-        }
-
-        animator: Animator{
-            highlight: {
-                default: @off
-                off: AnimatorState{
-                    redraw: true,
-                    from: { all: Forward {duration: 2.0} }
-                    ease: ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: { draw_bg: {highlight: 0.0} }
-                }
-                on: AnimatorState{
-                    redraw: true,
-                    from: { all: Forward {duration: 0.5} }
-                    ease: ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: { draw_bg: {highlight: 1.0} }
-                }
-            }
-            hover: {
-                default: @off
-                off: AnimatorState{
-                    redraw: true,
-                    from: { all: Snap }
-                    apply: { draw_bg: {hover: 0.0} }
-                }
-                on: AnimatorState{
-                    redraw: true,
-                    from: { all: Snap }
-                    apply: { draw_bg: {hover: 1.0} }
-                }
-            }
-        }
-
-        // A preview of the earlier message that this message was in reply to.
-        replied_to_message := mod.widgets.RepliedToMessage {
-            flow: Right
-            margin: Inset{ bottom: 3, top: 10 }
-            replied_to_message_content +: {
-                margin +: { left: 29 }
-                padding +: { bottom: 10 }
-            }
-        }
-
-        body := View {
-            width: Fill,
-            height: Fit
-            flow: Right,
-            padding: Inset{top: 0, bottom: 10, left: 10, right: 10},
-
-            profile := View {
-                align: Align{x: 0.5, y: 0.0} // centered horizontally, top aligned
-                width: 65.0,
-                height: Fit,
-                margin: Inset{top: #(MESSAGE_PROFILE_TOP_MARGIN), right: 10}
-                flow: Down,
-                avatar := Avatar {
-                    width: #(MESSAGE_PROFILE_AVATAR_SIZE),
-                    height: #(MESSAGE_PROFILE_AVATAR_SIZE),
-                }
-                edited_indicator := EditedIndicator { }
-                tsp_sign_indicator := TspSignIndicator { }
-            }
-
-            content := View {
-                width: Fill,
-                height: Fit
-                flow: Down,
-                padding: 0.0
-
-                username_view := View {
-                    flow: Right,
-                    align: Align{y: 0.5},
-                    width: Fit,
-                    height: #(MESSAGE_USERNAME_ROW_HEIGHT),
-                    margin: Inset{
-                        top: #(MESSAGE_USERNAME_ROW_TOP_MARGIN),
-                        bottom: #(MESSAGE_USERNAME_ROW_BOTTOM_MARGIN),
-                    }
-                    username := Label {
-                        width: Fit,
-                        flow: Right, // do not wrap
-                        padding: 0,
-                        margin: Inset{right: #(MESSAGE_USERNAME_RIGHT_MARGIN)}
-                        max_lines: 1
-                        text_overflow: Ellipsis
-                        draw_text +: {
-                            text_style: USERNAME_TEXT_STYLE {},
-                            color: (USERNAME_TEXT_COLOR)
-                        }
-                        text: ""
-                    }
-                    bot_badge := RoundedView {
-                        visible: false
-                        width: Fit
-                        height: #(BOT_BADGE_HEIGHT)
-                        align: Align{x: 0.5, y: 0.5}
-                        padding: Inset{left: #(BOT_BADGE_HORIZONTAL_PADDING), right: #(BOT_BADGE_HORIZONTAL_PADDING)}
-                        show_bg: true
-                        draw_bg +: {
-                            color: (RBX_ACCENT_SOFT)
-                            border_radius: #(BOT_BADGE_BORDER_RADIUS)
-                        }
-                        bot_badge_label := Label {
-                            width: Fit
-                            height: Fit
-                            padding: 0
-                            draw_text +: {
-                                text_style: REGULAR_TEXT {
-                                    font_size: #(BOT_BADGE_TEXT_FONT_SIZE)
-                                    top_drop: #(BOT_BADGE_TEXT_TOP_DROP)
-                                }
-                                color: (RBX_ACCENT)
-                            }
-                            text: "bot"
-                        }
-                    }
-                    timestamp := Timestamp {
-                        margin: Inset{ left: (SPACE_XS) }
-                    }
-                }
-
-                bot_message_card := View {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    flow: Down
-                    spacing: 6.0
-                    margin: Inset{ top: 1.0, bottom: 3.0 }
-
-                    bot_status_strip := RoundedView {
-                        visible: false
-                        width: Fit
-                        height: Fit
-                        padding: Inset{ left: 10.0, right: 10.0, top: 5.0, bottom: 5.0 }
-                        show_bg: true
-                        draw_bg +: {
-                            color: (mod.widgets.COLOR_BOT_STATUS_BG)
-                            border_radius: 10.0
-                        }
-
-                        bot_status_label := Label {
-                            width: Fit
-                            height: Fit
-                            draw_text +: {
-                                text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                color: (mod.widgets.COLOR_BOT_STATUS_TEXT)
-                            }
-                            text: ""
-                        }
-                    }
-
-                    bot_body_card := RoundedView {
-                        width: Fill
-                        height: Fit
-                        flow: Down
-                        padding: Inset{ left: 14.0, right: 14.0, top: 12.0, bottom: 12.0 }
-                        show_bg: true
-                        draw_bg +: {
-                            color: (mod.widgets.COLOR_BOT_CARD_BG)
-                            border_radius: 6.0
-                            border_size: 1.0
-                            border_color: (mod.widgets.COLOR_BOT_CARD_BORDER)
-                        }
-
-                        bot_card_body := HtmlOrPlaintext { }
-                        bot_card_markdown := mod.widgets.BotTimelineMarkdown {
-                            body: ""
-                        }
-                        bot_card_markdown_plain := mod.widgets.BotTimelineMarkdown {
-                            use_code_block_widget: false
-                            body: ""
-                        }
-
-                        // Card footer: the fold affordance and the permalink sit
-                        // on one row so both stay reachable while folded.
-                        bot_card_footer_row := View {
-                            width: Fill
-                            height: Fit
-                            flow: Right
-                            align: Align{y: 0.5}
-                            spacing: 10.0
-                            margin: Inset{ top: 6.0 }
-
-                            // Fold affordance for long agent replies. Only shown
-                            // when the body exceeded the fold threshold.
-                            bot_body_fold_toggle := mod.widgets.SmallStateGroupToggleButton {
-                                visible: false
-                                padding: Inset{ left: 0.0, right: 6.0, top: 0.0, bottom: 0.0 }
-                                draw_text +: {
-                                    text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                    // The base widget only defines color/hover/down,
-                                    // so an unset focus colour turned the label white
-                                    // the moment a click focused it, until a later
-                                    // redraw happened to reset the state.
-                                    color: (mod.widgets.RBX_ACCENT)
-                                    color_hover: (mod.widgets.RBX_ACCENT_HOVER)
-                                    color_down: (mod.widgets.RBX_ACCENT_PRESSED)
-                                    color_focus: (mod.widgets.RBX_ACCENT)
-                                }
-                                text: ""
-                            }
-
-                            // The bridge's permalink, pinned here so folding the
-                            // body never hides it.
-                            bot_permalink_link := LinkLabel {
-                                visible: false
-                                width: Fit
-                                height: Fit
-                                padding: 0
-                                margin: 0
-                                spacing: 0
-                                draw_text +: {
-                                    text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                    // The base widget only defines color/hover/down,
-                                    // so an unset focus colour turned the label white
-                                    // the moment a click focused it, until a later
-                                    // redraw happened to reset the state.
-                                    color: (mod.widgets.RBX_ACCENT)
-                                    color_hover: (mod.widgets.RBX_ACCENT_HOVER)
-                                    color_down: (mod.widgets.RBX_ACCENT_PRESSED)
-                                    color_focus: (mod.widgets.RBX_ACCENT)
-                                }
-                                text: ""
-                            }
-                        }
-                    }
-
-                }
-
-                message := HtmlOrPlaintext { }
-                splash_card := Splash { }
-                action_buttons := View {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    flow: Down
-                    spacing: 6.0
-                    margin: Inset{ top: 8.0, bottom: 2.0 }
-
-                    approval_request_view := mod.widgets.AgentApprovalCard {}
-
-                    action_button_row := View {
-                        visible: false
-                        width: Fill
-                        height: Fit
-                        flow: Flow.Right{wrap: true}
-                        spacing: 8.0
-
-                        action_button_slot_0 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_1 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_2 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_3 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_4 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_5 := mod.widgets.MessageActionButtonSlot {}
-                    }
-                }
-                link_preview_view := mod.widgets.LinkPreview {}
-                download_section := mod.widgets.MessageDownloadSection {}
-                send_failure_section := mod.widgets.SendFailureSection {}
-                message_action_bar := mod.widgets.MessageMetaBand {}
-                View {
-                    width: Fill,
-                    height: Fit
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-        }
-    }
-
-    // The view used for a condensed message that came right after another message
-    // from the same sender, and thus doesn't need to display the sender's profile again.
-    mod.widgets.CondensedMessage = mod.widgets.Message {
-        padding: Inset{ top: 2.0, bottom: 2.0 }
-        replied_to_message +: {
-            replied_to_message_content +: {
-                margin: Inset{ left: 74, bottom: 5.0 }
-            }
-        }
-        body := View {
-            width: Fill,
-            height: Fit
-            flow: Right,
-            padding: Inset{ top: 0, bottom: 2.5, left: 10.0, right: 10.0 },
-            profile := View {
-                align: Align{x: 0.5, y: 0.0} // centered horizontally, top aligned
-                width: 65.0,
-                height: Fit,
-                flow: Down,
-                timestamp := Timestamp {
-                    margin: Inset{top: 2.5}
-                }
-                edited_indicator := EditedIndicator { }
-                tsp_sign_indicator := TspSignIndicator { }
-            }
-            content := View {
-                width: Fill,
-                height: Fit,
-                flow: Down,
-                padding: Inset{ left: 10.0 }
-
-                bot_message_card := View {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    flow: Down
-                    spacing: 6.0
-                    margin: Inset{ top: 1.0, bottom: 3.0 }
-
-                    bot_status_strip := RoundedView {
-                        visible: false
-                        width: Fit
-                        height: Fit
-                        padding: Inset{ left: 10.0, right: 10.0, top: 5.0, bottom: 5.0 }
-                        show_bg: true
-                        draw_bg +: {
-                            color: (mod.widgets.COLOR_BOT_STATUS_BG)
-                            border_radius: 10.0
-                        }
-
-                        bot_status_label := Label {
-                            width: Fit
-                            height: Fit
-                            draw_text +: {
-                                text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                color: (mod.widgets.COLOR_BOT_STATUS_TEXT)
-                            }
-                            text: ""
-                        }
-                    }
-
-                    bot_body_card := RoundedView {
-                        width: Fill
-                        height: Fit
-                        flow: Down
-                        padding: Inset{ left: 14.0, right: 14.0, top: 12.0, bottom: 12.0 }
-                        show_bg: true
-                        draw_bg +: {
-                            color: (mod.widgets.COLOR_BOT_CARD_BG)
-                            border_radius: 6.0
-                            border_size: 1.0
-                            border_color: (mod.widgets.COLOR_BOT_CARD_BORDER)
-                        }
-
-                        bot_card_body := HtmlOrPlaintext { }
-                        bot_card_markdown := mod.widgets.BotTimelineMarkdown {
-                            body: ""
-                        }
-                        bot_card_markdown_plain := mod.widgets.BotTimelineMarkdown {
-                            use_code_block_widget: false
-                            body: ""
-                        }
-
-                        // Card footer: the fold affordance and the permalink sit
-                        // on one row so both stay reachable while folded.
-                        bot_card_footer_row := View {
-                            width: Fill
-                            height: Fit
-                            flow: Right
-                            align: Align{y: 0.5}
-                            spacing: 10.0
-                            margin: Inset{ top: 6.0 }
-
-                            // Fold affordance for long agent replies. Only shown
-                            // when the body exceeded the fold threshold.
-                            bot_body_fold_toggle := mod.widgets.SmallStateGroupToggleButton {
-                                visible: false
-                                padding: Inset{ left: 0.0, right: 6.0, top: 0.0, bottom: 0.0 }
-                                draw_text +: {
-                                    text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                    // The base widget only defines color/hover/down,
-                                    // so an unset focus colour turned the label white
-                                    // the moment a click focused it, until a later
-                                    // redraw happened to reset the state.
-                                    color: (mod.widgets.RBX_ACCENT)
-                                    color_hover: (mod.widgets.RBX_ACCENT_HOVER)
-                                    color_down: (mod.widgets.RBX_ACCENT_PRESSED)
-                                    color_focus: (mod.widgets.RBX_ACCENT)
-                                }
-                                text: ""
-                            }
-
-                            // The bridge's permalink, pinned here so folding the
-                            // body never hides it.
-                            bot_permalink_link := LinkLabel {
-                                visible: false
-                                width: Fit
-                                height: Fit
-                                padding: 0
-                                margin: 0
-                                spacing: 0
-                                draw_text +: {
-                                    text_style: mod.widgets.MESSAGE_TEXT_STYLE { font_size: 9.5 }
-                                    // The base widget only defines color/hover/down,
-                                    // so an unset focus colour turned the label white
-                                    // the moment a click focused it, until a later
-                                    // redraw happened to reset the state.
-                                    color: (mod.widgets.RBX_ACCENT)
-                                    color_hover: (mod.widgets.RBX_ACCENT_HOVER)
-                                    color_down: (mod.widgets.RBX_ACCENT_PRESSED)
-                                    color_focus: (mod.widgets.RBX_ACCENT)
-                                }
-                                text: ""
-                            }
-                        }
-                    }
-
-                }
-
-                message := HtmlOrPlaintext { }
-                action_buttons := View {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    flow: Down
-                    spacing: 6.0
-                    margin: Inset{ top: 8.0, bottom: 2.0 }
-
-                    approval_request_view := mod.widgets.AgentApprovalCard {}
-
-                    action_button_row := View {
-                        visible: false
-                        width: Fill
-                        height: Fit
-                        flow: Flow.Right{wrap: true}
-                        spacing: 8.0
-
-                        action_button_slot_0 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_1 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_2 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_3 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_4 := mod.widgets.MessageActionButtonSlot {}
-                        action_button_slot_5 := mod.widgets.MessageActionButtonSlot {}
-                    }
-                }
-                link_preview_view := mod.widgets.LinkPreview {}
-                download_section := mod.widgets.MessageDownloadSection {}
-                send_failure_section := mod.widgets.SendFailureSection {}
-                message_action_bar := mod.widgets.MessageMetaBand {}
-                View {
-                    width: Fill,
-                    height: Fit
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-        }
-    }
-
-    mod.widgets.IMG_MSG_FIT = Fit{max: FitBound.Abs(200.0)}
-    mod.widgets.STICKER_HEIGHT = 150.0
-
-    // Sticker message templates: fixed height, width determined by aspect ratio.
-    mod.widgets.StickerMessage = mod.widgets.Message {
-        body +: {
-            content +: {
-                width: Fill,
-                height: Fit
-                padding: Inset{ left: 10.0 }
-
-                message := TextOrImage {
-                    width: Fit, height: Fit,
-                    image_view +: { width: Fit, height: Fit, image +: {
-                        height: (mod.widgets.STICKER_HEIGHT)
-                        width: (mod.widgets.STICKER_HEIGHT)
-                        fit: ImageFit.Smallest
-                    } }
-                    default_image_view +: { width: Fit, height: Fit, image +: {
-                        height: (mod.widgets.STICKER_HEIGHT)
-                        width: (mod.widgets.STICKER_HEIGHT)
-                        fit: ImageFit.Smallest
-                    } }
-                }
-                View {
-                    width: Fill,
-                    height: Fit,
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-        }
-    }
-
-    mod.widgets.CondensedStickerMessage = mod.widgets.CondensedMessage {
-        body +: {
-            content +: {
-                message := TextOrImage {
-                    width: Fit, height: Fit,
-                    image_view +: { width: Fit, height: Fit, image +: {
-                        height: (mod.widgets.STICKER_HEIGHT)
-                        width: (mod.widgets.STICKER_HEIGHT)
-                        fit: ImageFit.Smallest
-                    } }
-                    default_image_view +: { width: Fit, height: Fit, image +: {
-                        height: (mod.widgets.STICKER_HEIGHT)
-                        width: (mod.widgets.STICKER_HEIGHT)
-                        fit: ImageFit.Smallest
-                    } }
-                }
-                View {
-                    width: Fill,
-                    height: Fit,
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-        }
-    }
-
-    // The view used for each static image-based message event in a room's timeline.
-    // This excludes stickers and other animated GIFs, video clips, audio clips, etc.
-    mod.widgets.ImageMessage = mod.widgets.Message {
-        body +: {
-            content +: {
-                width: Fill,
-                height: Fit
-                padding: Inset{ left: 10.0 }
-
-                message := TextOrImage {
-                    image_view +: { image +: {
-                        height: (mod.widgets.IMG_MSG_FIT)
-                    } }
-                    default_image_view +: { image +: {
-                        height: (mod.widgets.IMG_MSG_FIT)
-                    } }
-                }
-                download_section := mod.widgets.MessageDownloadSection {}
-                send_failure_section := mod.widgets.SendFailureSection {}
-                animated_message := mod.widgets.AnimatedImage { visible: false }
-                View {
-                    width: Fill,
-                    height: Fit,
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-
-        }
-    }
-
-    // The view used for a condensed image message that came right after another message
-    // from the same sender, and thus doesn't need to display the sender's profile again.
-    // This excludes stickers and other animated GIFs, video clips, audio clips, etc.
-    mod.widgets.CondensedImageMessage = mod.widgets.CondensedMessage {
-        body +: {
-            content +: {
-                message := TextOrImage {
-                    image_view +: { image +: {
-                        height: (mod.widgets.IMG_MSG_FIT)
-                    } }
-                    default_image_view +: { image +: {
-                        height: (mod.widgets.IMG_MSG_FIT)
-                    } }
-                }
-                download_section := mod.widgets.MessageDownloadSection {}
-                send_failure_section := mod.widgets.SendFailureSection {}
-                animated_message := mod.widgets.AnimatedImage { visible: false }
-                View {
-                    width: Fill,
-                    height: Fit,
-                    flow: Right,
-                    reaction_list := mod.widgets.ReactionList { }
-                }
-                thread_root_summary := mod.widgets.ThreadRootSummary {}
-            }
-        }
-    }
-
-    // Video message template. Embeds the inline `VideoMessagePlayer` widget
-    // above the existing html caption/metadata. The player is populated by
-    // `populate_video_message_content` in this file.
-    mod.widgets.VideoMessage = mod.widgets.Message {
-        body +: {
-            content +: {
-                video_player := mod.widgets.VideoMessagePlayer {}
-            }
-        }
-    }
-
-    mod.widgets.AudioMessage = mod.widgets.Message {
-        body +: {
-            content +: {
-                audio_player := mod.widgets.AudioMessagePlayer {}
-            }
-        }
-    }
-
-
-    // The view used for each state event (non-messages) in a room's timeline.
-    // The timestamp, profile picture, and text are all very small.
-    mod.widgets.SmallStateEvent = View {
-        width: Fill,
-        height: Fit,
-        flow: Right,
-        margin: Inset{ top: 4.0, bottom: 4.0}
-        padding: Inset{ top: 1.0, bottom: 1.0, right: 10.0 }
-        spacing: 0.0
-        cursor: MouseCursor.Default
-
-        body := View {
-            width: Fill,
-            height: Fit
-            flow: Down,
-            padding: Inset{ left: 7.0, top: 2.0, bottom: 2.0 }
-            spacing: 4.0
-
-            group_header := View {
-                visible: false
-                width: Fill,
-                height: Fit
-                flow: Right
-                spacing: 0.0
-                // 43 + the parent `body`'s 7 = the 50 that
-                // `SmallStateEventsSummary` uses, so the summary text keeps its
-                // left edge when the group is expanded instead of shifting 7px.
-                padding: Inset{ left: 43.0, right: 10.0, bottom: 1.0 }
-
-                group_summary_label := Label {
-                    width: Fit,
-                    height: Fit
-                    draw_text +: {
-                        text_style: SMALL_STATE_TEXT_STYLE {}
-                        color: (SMALL_STATE_TEXT_COLOR)
-                    }
-                    text: ""
-                }
-
-                spacer := View {
-                    width: Fill
-                    height: Fit
-                }
-
-                state_group_toggle_button := mod.widgets.SmallStateGroupToggleButton {
-                    width: Fit
-                    height: Fit
-                    margin: Inset{ top: 1.0 }
-                    text: ""
-                }
-            }
-
-            event_row := View {
-                width: Fill,
-                height: Fit
-                flow: Right,
-                spacing: 5.0
-
-                left_container := View {
-                    align: Align{x: 0.5, y: 0}
-                    width: 70.0,
-                    height: Fit
-
-                    timestamp := Timestamp {
-                        margin: Inset{top: 3}
-                    }
-                }
-
-                avatar := Avatar {
-                    width: 19.,
-                    height: 19.,
-                    margin: 0
-
-                    text_view +: {
-                        text +: {
-                            draw_text +: {
-                                text_style: TITLE_TEXT { font_size: 7.0 }
-                            }
-                        }
-                    }
-                }
-
-                // Show an invite button only for a `Knocked` room membership change.
-                // All other small state events will not show this button.
-                invite_user_button := RobrixPositiveIconButton {
-                    visible: false
-                    margin: Inset{ top: -1.5, left: 2, right: 2}
-                    padding: Inset{top: 4, bottom: 4, left: 9, right: 9}
-                    draw_bg +: {
-                        border_size: 0.75
-                    }
-                    draw_icon.svg: (ICON_ADD_USER)
-                    draw_text.text_style: SMALL_STATE_TEXT_STYLE {}
-                    icon_walk: Walk{width: 15, height: Fit, margin: Inset{right: -4}}
-                    text: ""
-                }
-
-                content := Label {
-                    width: Fill,
-                    height: Fit
-                    flow: Flow.Right{wrap: true},
-                    margin: Inset{top: 2.5}
-                    padding: Inset{ top: 0.0, bottom: 0.0, left: 0.0, right: 0.0 }
-                    draw_text +: {
-                        text_style: SMALL_STATE_TEXT_STYLE {},
-                        color: (SMALL_STATE_TEXT_COLOR)
-                    }
-                    text: ""
-                }
-
-                avatar_row := mod.widgets.AvatarRow {}
-            }
-        }
-    }
-
-    // The summary row shown for a collapsed group of adjacent small state events.
-    mod.widgets.SmallStateEventsSummary = View {
-        width: Fill,
-        height: Fit,
-        flow: Right,
-        margin: Inset{ top: 4.0, bottom: 4.0}
-        padding: Inset{ left: 50.0, top: 1.0, bottom: 1.0, right: 10.0 }
-        spacing: 7.0
-        cursor: MouseCursor.Default
-
-        summary_label := Label {
-            width: Fit,
-            height: Fit
-            flow: Right
-            margin: Inset{top: 1.5}
-            draw_text +: {
-                text_style: SMALL_STATE_TEXT_STYLE {}
-                color: (SMALL_STATE_TEXT_COLOR)
-            }
-            text: ""
-        }
-
-        spacer := View {
-            width: Fill
-            height: Fit
-        }
-
-        state_group_toggle_button := mod.widgets.SmallStateGroupToggleButton {
-            width: Fit
-            height: Fit
-            margin: Inset{ left: 2.0, top: 1.0 }
-            text: ""
-        }
-    }
-
-
-    // The view used for each day divider in a room's timeline.
-    // The date text is centered between two horizontal lines.
-    mod.widgets.DateDivider = View {
-        width: Fill,
-        height: Fit,
-        margin: Inset{top: 7.0, bottom: 7.0}
-        flow: Right,
-        padding: Inset{left: 7.0, right: 7.0},
-        spacing: 0.0,
-        align: Align{x: 0.5, y: 0.5} // center horizontally and vertically
-
-        left_line := LineH { }
-
-        date := Label {
-            padding: Inset{left: 7.0, right: 7.0}
-            draw_text +: {
-                text_style: TEXT_SUB {},
-                color: (COLOR_DIVIDER_DARK)
-            }
-            text: ""
-        }
-
-        right_line := LineH { }
-    }
-
-    // The view used for the divider indicating where the user's last-viewed message is.
-    // This is implemented as a DateDivider with a different color and a fixed text label.
-    mod.widgets.ReadMarker = mod.widgets.DateDivider {
-        left_line := LineH {
-            draw_bg.color: (mod.widgets.COLOR_READ_MARKER)
-        }
-
-        date := Label {
-            draw_text.color: (mod.widgets.COLOR_READ_MARKER)
-            text: ""
-        }
-
-        right_line := LineH {
-            draw_bg.color: (mod.widgets.COLOR_READ_MARKER)
-        }
-    }
-
-
-    // The top space is used to display a loading message while the room is being paginated.
-    mod.widgets.TopSpace = SolidView {
-        visible: false,
-        width: Fill,
-        height: Fit,
-        align: Align{x: 0.5, y: 0}
-        flow: Right,
-        show_bg: true,
-        draw_bg.color: #xDAF5E5F0, // mostly opaque light green
-
-        label := Label {
-            width: Fill,
-            height: Fit,
-            align: Align{x: 0.5, y: 0.5},
-            flow: Right,
-            padding: Inset{ top: 10.0, bottom: 7.0, left: 15.0, right: 15.0 }
-            draw_text +: {
-                text_style: MESSAGE_TEXT_STYLE { font_size: 10 },
-                color: (TIMESTAMP_TEXT_COLOR)
-            }
-            text: ""
-        }
-    }
-
-    mod.widgets.ThreadsPaneEntry = #(ThreadsPaneEntry::register_widget(vm)) {
-        ..mod.widgets.RoundedView
-
-        width: Fill
-        height: Fit
-        flow: Down
-        spacing: 5
-        padding: Inset{top: 12, right: 12, bottom: 12, left: 12}
-        margin: Inset{left: 12, right: 12, top: 6, bottom: 0}
-        cursor: MouseCursor.Hand
-
-        show_bg: true
-        draw_bg +: {
-            color: #F8FAFD
-            border_radius: 4.0
-            border_size: 1.0
-            border_color: #D8E0EA
-        }
-
-        title_row := View {
-            width: Fill
-            height: Fit
-            flow: Right
-            spacing: 8
-
-            title := HtmlOrPlaintext {
-                width: Fill
-                height: Fit
-            }
-
-            time := Label {
-                width: Fit
-                height: Fit
-                draw_text +: {
-                    text_style: TIMESTAMP_TEXT_STYLE { font_size: 7.5 }
-                    color: (TIMESTAMP_TEXT_COLOR)
-                }
-                text: ""
-            }
-        }
-
-        subtitle := Label {
-            width: Fill
-            height: Fit
-            flow: Flow.Right{wrap: true}
-            draw_text +: {
-                text_style: MESSAGE_TEXT_STYLE { font_size: 9.8 }
-                color: #7B7B7B
-            }
-            text: ""
-        }
-
-        preview := HtmlOrPlaintext {
-            width: Fill
-            height: Fit
-        }
-    }
-
-    // Floating circular button that opens the `ThreadsSlidingPane`.
-    // Mirrors `SearchMessagesButton`'s layout (Fill/Fill overlay aligned
-    // top-right) but reserves 96px on the right so it sits to the left of the
-    // search (48px) and info (rightmost) buttons.
-    mod.widgets.ThreadsButton = #(ThreadsButton::register_widget(vm)) {
-        width: Fill,
-        height: Fill,
-        flow: Overlay,
-        align: Align{x: 1.0, y: 0.0},
-        padding: Inset{right: 96},
-        visible: true,
-
-        View {
-            width: 65, height: 65,
-            align: Align{x: 0.5, y: 0.0},
-            flow: Overlay,
-
-            inner_button := RobrixIconButton {
-                spacing: 0,
-                width: 40, height: 40,
-                align: Align{x: 0.5, y: 0.5},
-                margin: Inset{top: 8},
-
-                draw_icon +: {
-                    svg: (ICON_THREADS),
-                    color: #555,
-                }
-                icon_walk: Walk{width: 18, height: 18}
-
-                draw_bg +: {
-                    background_color: #edededce,
-                    background_color_hover: #d0d0d0ce,
-                    pixel: fn() {
-                        let sdf = Sdf2d.viewport(self.pos * self.rect_size);
-                        let c = self.rect_size * 0.5;
-                        sdf.circle(c.x, c.x, c.x);
-                        sdf.fill_keep(mix(self.background_color, self.background_color_hover, self.hover));
-                        return sdf.result
-                    }
-                }
-            }
-        }
-    }
-
-    // Floating circular button that opens the `RoomInfoSlidingPane` (desktop
-    // only — mobile reaches room info via the RoomTopBar "Info" tab). Sits at
-    // the rightmost slot (no right inset); the search (48px) and threads (96px)
-    // buttons sit to its left.
-    mod.widgets.InfoButton = #(InfoButton::register_widget(vm)) {
-        width: Fill,
-        height: Fill,
-        flow: Overlay,
-        align: Align{x: 1.0, y: 0.0},
-        visible: true,
-
-        View {
-            width: 65, height: 65,
-            align: Align{x: 0.5, y: 0.0},
-            flow: Overlay,
-
-            inner_button := RobrixIconButton {
-                spacing: 0,
-                width: 40, height: 40,
-                align: Align{x: 0.5, y: 0.5},
-                margin: Inset{top: 8},
-
-                draw_icon +: {
-                    svg: (ICON_INFO),
-                    color: #555,
-                }
-                icon_walk: Walk{width: 18, height: 18}
-
-                draw_bg +: {
-                    background_color: #edededce,
-                    background_color_hover: #d0d0d0ce,
-                    pixel: fn() {
-                        let sdf = Sdf2d.viewport(self.pos * self.rect_size);
-                        let c = self.rect_size * 0.5;
-                        sdf.circle(c.x, c.x, c.x);
-                        sdf.fill_keep(mix(self.background_color, self.background_color_hover, self.hover));
-                        return sdf.result
-                    }
-                }
-            }
-        }
-    }
-
-    mod.widgets.ThreadsSlidingPane = #(ThreadsSlidingPane::register_widget(vm)) {
-        visible: false,
-        flow: Overlay,
-        width: Fill,
-        height: Fill,
-        align: Align{x: 1.0, y: 0}
-
-        bg_view := SolidView {
-            width: Fill
-            height: Fill
-            visible: false,
-            show_bg: true
-            draw_bg.color: (RBX_SCRIM)
-        }
-
-        main_content := SolidView {
-            width: 320,
-            height: Fill
-            flow: Down,
-            align: Align{x: 1.0}
-
-            show_bg: true,
-            draw_bg.color: (COLOR_PRIMARY)
-
-            header := View {
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{y: 0.5}
-                padding: Inset{top: 12, right: 10, bottom: 12, left: 15}
-
-                title := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: {
-                        text_style: USERNAME_TEXT_STYLE { font_size: 12.5 }
-                        color: #000
-                    }
-                    text: "Threads"
-                }
-
-                spacer := View {
-                    width: Fill
-                    height: Fit
-                }
-
-                close_button := RobrixNeutralIconButton {
-                    width: Fit,
-                    height: Fit,
-                    spacing: 0,
-                    padding: 15,
-                    draw_icon.svg: (ICON_CLOSE)
-                    icon_walk: Walk{width: 14, height: 14}
-                    text: ""
-                }
-            }
-
-            room_name := Label {
-                width: Fill
-                height: Fit
-                flow: Flow.Right{wrap: true}
-                padding: Inset{left: 15, right: 15, bottom: 10}
-                draw_text +: {
-                    text_style: MESSAGE_TEXT_STYLE { font_size: 10.5 }
-                    color: #6E6E6E
-                }
-                text: ""
-            }
-
-            loading_indicator := View {
-                visible: false
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{y: 0.5}
-                spacing: 8
-                padding: Inset{left: 15, right: 15, top: 6, bottom: 10}
-
-                spinner := LoadingSpinner {
-                    width: 18
-                    height: 18
-                }
-
-                loading_label := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: {
-                        text_style: MESSAGE_TEXT_STYLE { font_size: 10.5 }
-                        color: #7B7B7B
-                    }
-                    text: "Loading threads..."
-                }
-            }
-
-            empty_state := Label {
-                visible: false
-                width: Fill
-                height: Fit
-                flow: Flow.Right{wrap: true}
-                padding: Inset{left: 15, right: 15, top: 20, bottom: 20}
-                draw_text +: {
-                    text_style: MESSAGE_TEXT_STYLE { font_size: 10.5 }
-                    color: #7B7B7B
-                }
-                text: "No threads yet."
-            }
-
-            threads_list := PortalList {
-                width: Fill
-                height: Fill
-                flow: Down
-                max_pull_down: 0.0
-
-                ThreadEntry := mod.widgets.ThreadsPaneEntry {}
-            }
-        }
-
-        slide: 1.0,
-
-        animator: Animator {
-            panel: {
-                default: @hide
-                show: AnimatorState{
-                    redraw: true,
-                    from: {all: Forward {duration: 0.5}}
-                    ease: Ease.ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {
-                        slide: 0.0
-                    }
-                }
-                hide: AnimatorState{
-                    redraw: true,
-                    from: {all: Forward {duration: 0.5}}
-                    ease: Ease.ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {
-                        slide: 1.0
-                    }
-                }
-            }
-        }
-    }
-
-    mod.widgets.RoomInfoPeopleEntry = #(RoomInfoPeopleEntry::register_widget(vm)) {
-        width: Fill
-        height: Fit
-        flow: Down
-        cursor: MouseCursor.Hand
-
-        row := View {
-            width: Fill
-            height: Fit
-            flow: Right
-            align: Align{y: 0.5}
-            spacing: 12
-            padding: Inset{left: 14, right: 14, top: 11, bottom: 11}
-
-            avatar := Avatar {
-                width: 38
-                height: 38
-            }
-
-            name_wrap := View {
-                width: Fill
-                height: Fit
-                flow: Flow.Right{wrap: true}
-                align: Align{y: 0.5}
-
-                display_name := Label {
-                    width: Fit
-                    height: Fit
-                    flow: Flow.Right{wrap: true}
-                    // Gap to the badge lives here (not as container `spacing`)
-                    // so a wrapped bot_badge starts flush at the row's left
-                    // edge instead of inheriting an extra leading offset from
-                    // the wrap-flow container's spacing bookkeeping.
-                    margin: Inset{right: 6}
-                    draw_text +: {
-                        text_style: RBX_TEXT_BODY_STRONG {}
-                        color: (RBX_FG_PRIMARY)
-                    }
-                    text: ""
-                }
-
-                // Shown only for bot members (in place of the old " [bot]" suffix).
-                bot_badge := RoundedView {
-                    visible: false
-                    width: Fit
-                    height: Fit
-                    align: Align{x: 0.5, y: 0.5}
-                    padding: Inset{left: 7, right: 7, top: 2, bottom: 2}
-                    show_bg: true
-                    draw_bg +: { color: (RBX_ACCENT_SOFT), border_radius: (RBX_RADIUS_PILL) }
-                    Label {
-                        width: Fit, height: Fit
-                        draw_text +: { text_style: RBX_TEXT_BADGE {}, color: (RBX_ACCENT) }
-                        text: "bot"
-                    }
-                }
-            }
-
-            // Role chip (Creator / Admin / Moderator) — hidden for plain members.
-            level_chip := RoundedView {
-                visible: false
-                width: Fit
-                height: Fit
-                align: Align{x: 0.5, y: 0.5}
-                padding: Inset{left: 9, right: 9, top: 3, bottom: 3}
-                show_bg: true
-                draw_bg +: { color: (RBX_INFO_BG), border_radius: (RBX_RADIUS_PILL) }
-                level := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: { text_style: RBX_TEXT_BADGE {}, color: (RBX_INFO_FG) }
-                    text: ""
-                }
-            }
-        }
-
-        member_divider := RoundedView {
-            width: Fill
-            height: 1.0
-            margin: Inset{left: 64}
-            show_bg: true
-            draw_bg +: { color: (RBX_STROKE_SOFT) }
-        }
-    }
-
-    mod.widgets.RoomInfoSlidingPane = #(RoomInfoSlidingPane::register_widget(vm)) {
-        visible: false,
-        flow: Overlay,
-        width: Fill,
-        height: Fill,
-        align: Align{x: 1.0, y: 0}
-
-        bg_view := SolidView {
-            width: Fill
-            height: Fill
-            visible: false,
-            show_bg: true
-            draw_bg.color: (RBX_SCRIM)
-        }
-
-        main_content := SolidView {
-            width: 320,
-            height: Fill
-            flow: Down,
-            align: Align{x: 1.0}
-
-            show_bg: true,
-            // Cool off-white page so the white cards below read as distinct
-            // surfaces (the grouped-list look from the spec).
-            draw_bg.color: (RBX_BG_CANVAS)
-
-            header := View {
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{y: 0.5}
-                padding: Inset{top: 12, right: 10, bottom: 12, left: 15}
-
-                back_button := RobrixNeutralIconButton {
-                    visible: false
-                    width: Fit,
-                    height: Fit,
-                    spacing: 0,
-                    // Left inset matches the room top bar's back icon (header pad
-                    // 6 + button pad 6 = 12) so the two arrows line up vertically.
-                    padding: Inset{left: 12, right: 3, top: 8, bottom: 8}
-                    draw_bg +: {
-                        color: #0000
-                        color_hover: (RBX_BG_HOVER)
-                        color_down: (RBX_BG_PRESSED)
-                        border_size: 0.0
-                        border_color: #0000
-                        border_color_hover: #0000
-                        border_color_down: #0000
-                        border_radius: (RBX_RADIUS_XS)
-                    }
-                    draw_icon +: { svg: (ICON_JUMP), color: (RBX_FG_SECONDARY) }
-                    icon_walk: Walk{width: 16, height: 16}
-                    text: ""
-                }
-
-                // People sub-page count, shown inline with the back button so the
-                // two are horizontally aligned (also visible in the inline tab,
-                // where `title` is hidden).
-                members_header_count := Label {
-                    visible: false
-                    width: Fit
-                    height: Fit
-                    margin: Inset{left: 0}
-                    draw_text +: {
-                        text_style: RBX_TEXT_SECTION_TITLE {}
-                        color: (RBX_FG_PRIMARY)
-                    }
-                    text: ""
-                }
-
-                title := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: {
-                        text_style: USERNAME_TEXT_STYLE { font_size: 12.5 }
-                        color: #000
-                    }
-                    text: "Info"
-                }
-
-                spacer := View {
-                    width: Fill
-                    height: Fit
-                }
-
-                close_button := RobrixNeutralIconButton {
-                    width: Fit,
-                    height: Fit,
-                    spacing: 0,
-                    padding: 15,
-                    draw_icon.svg: (ICON_CLOSE)
-                    icon_walk: Walk{width: 14, height: 14}
-                    text: ""
-                }
-            }
-
-            content_scroll := ScrollYView {
-                width: Fill
-                height: Fill
-                flow: Down
-
-                info_view := View {
-                    width: Fill
-                    height: Fit
-                    flow: Down
-                    spacing: 12
-                    padding: Inset{left: 12, right: 12, top: 12, bottom: 12}
-
-                    // ===== Hero / summary card =====
-                    // Hero sits directly on the page (no card chrome) per spec.
-                    summary_card := View {
-                        width: Fill
-                        height: Fit
-                        flow: Down
-                        spacing: 12
-                        padding: Inset{left: 14, right: 14, top: 4, bottom: 4}
-
-                        hero_row := View {
-                            width: Fill
-                            height: Fit
-                            flow: Right
-                            spacing: 13
-                            align: Align{y: 0.0}
-
-                            room_avatar := Avatar {
-                                width: 56
-                                height: 56
-                            }
-
-                            room_meta := View {
-                                width: Fill
-                                height: Fit
-                                flow: Down
-                                spacing: 2
-
-                                name_row := View {
-                                    width: Fill
-                                    height: Fit
-                                    flow: Right
-                                    align: Align{y: 0.5}
-                                    spacing: 6
-
-                                    // Room name + bot pill live together in a Fill
-                                    // sub-row so the pill trails the (capped,
-                                    // ellipsized) name text directly, the same way
-                                    // the rooms list packs its bot pill snug against
-                                    // the name instead of at the row's far edge.
-                                    // This sub-row keeps the same Fill role that
-                                    // room_name_value used to play here, so
-                                    // favorite_button's pinned-right position is
-                                    // unaffected.
-                                    title_wrap := View {
-                                        width: Fill
-                                        height: Fit
-                                        flow: Right
-                                        align: Align{y: 0.5}
-                                        spacing: 6
-
-                                        room_name_value := Label {
-                                            width: Fit{max: FitBound.Rel{base: Base.Full, factor: 0.82}}
-                                            height: Fit
-                                            flow: Flow.Right{wrap: false}
-                                            max_lines: 1
-                                            text_overflow: Ellipsis
-                                            draw_text +: {
-                                                text_style: RBX_TEXT_SECTION_TITLE {}
-                                                color: (RBX_FG_PRIMARY)
-                                            }
-                                            text: ""
-                                        }
-
-                                        // Bot indicator pill, styled to match the
-                                        // rooms-list / timeline bot pill exactly.
-                                        title_bot_pill := RoundedView {
-                                            visible: false
-                                            width: Fit
-                                            height: 16.0
-                                            align: Align{x: 0.5, y: 0.5}
-                                            padding: Inset{left: 6.0, right: 6.0}
-                                            show_bg: true
-                                            new_batch: true
-                                            draw_bg +: {
-                                                color: (RBX_ACCENT_SOFT)
-                                                border_radius: 3.0
-                                            }
-                                            Label {
-                                                width: Fit, height: Fit, padding: 0
-                                                draw_text +: {
-                                                    text_style: REGULAR_TEXT { font_size: 8.5, top_drop: -0.08 }
-                                                    color: (RBX_ACCENT)
-                                                }
-                                                text: "bot"
-                                            }
-                                        }
-                                    }
-
-                                    favorite_button := View {
-                                        width: Fit
-                                        height: Fit
-                                        flow: Overlay
-                                        align: Align{x: 0.5, y: 0.5}
-                                        padding: 1
-                                        cursor: MouseCursor.Hand
-
-                                        star_outline := View {
-                                            width: Fit
-                                            height: Fit
-                                            Icon {
-                                                width: 17
-                                                height: 17
-                                                align: Align{x: 0.5, y: 0.5}
-                                                draw_icon +: {
-                                                    svg: (ICON_STAR)
-                                                    color: (RBX_FG_TERTIARY)
-                                                }
-                                                icon_walk: Walk{width: 17, height: 17}
-                                            }
-                                        }
-
-                                        star_filled := View {
-                                            visible: false
-                                            width: Fit
-                                            height: Fit
-                                            Icon {
-                                                width: 17
-                                                height: 17
-                                                align: Align{x: 0.5, y: 0.5}
-                                                draw_icon +: {
-                                                    svg: (ICON_STAR_FILLED)
-                                                    color: (RBX_WARNING_FG)
-                                                }
-                                                icon_walk: Walk{width: 17, height: 17}
-                                            }
-                                        }
-                                    }
-                                }
-
-                                room_id_row := View {
-                                    width: Fill
-                                    height: Fit
-                                    flow: Right
-                                    align: Align{y: 0.5}
-                                    spacing: 4
-
-                                    room_id_value := Label {
-                                        width: Fill
-                                        height: Fit
-                                        flow: Flow.Right{wrap: true}
-                                        draw_text +: {
-                                            text_style: RBX_TEXT_META {}
-                                            color: (RBX_FG_TERTIARY)
-                                        }
-                                        text: ""
-                                    }
-
-                                    copy_room_id_button := RobrixNeutralIconButton {
-                                        width: Fit
-                                        height: Fit
-                                        padding: 2
-                                        spacing: 0
-                                        margin: 0
-                                        draw_bg +: {
-                                            color: #00000000
-                                            color_hover: #00000000
-                                            color_down: #00000000
-                                            border_size: 0.0
-                                        }
-                                        draw_icon +: { svg: (ICON_COPY), color: (RBX_FG_TERTIARY) }
-                                        icon_walk: Walk{width: 13, height: 13}
-                                        text: ""
-                                    }
-                                }
-
-                                // visibility · members · encryption — left-aligned
-                                // directly under the room id. Wraps to the next
-                                // line when the meta area (constrained by the
-                                // avatar on the left) is too narrow to fit all
-                                // three on one row, so the encryption item is no
-                                // longer clipped at the right edge.
-                                meta_row := View {
-                                    width: Fill
-                                    height: Fit
-                                    flow: Flow.Right{wrap: true}
-                                    spacing: 12
-                                    wrap_spacing: 6
-                                    align: Align{y: 0.5}
-                                    margin: Inset{top: 3}
-
-                                    visibility_meta := View {
-                                        width: Fit
-                                        height: Fit
-                                        flow: Right
-                                        spacing: 4
-                                        align: Align{y: 0.5}
-
-                                        Icon {
-                                            width: 13
-                                            height: 13
-                                            draw_icon +: { svg: (ICON_GLOBE), color: (RBX_FG_SECONDARY) }
-                                            icon_walk: Walk{width: 13, height: 13}
-                                        }
-                                        visibility_value := Label {
-                                            width: Fit
-                                            height: Fit
-                                            draw_text +: { text_style: RBX_TEXT_META {}, color: (RBX_FG_SECONDARY) }
-                                            text: ""
-                                        }
-                                    }
-
-                                    members_meta := View {
-                                        width: Fit
-                                        height: Fit
-                                        flow: Right
-                                        spacing: 4
-                                        align: Align{y: 0.5}
-
-                                        Icon {
-                                            width: 13
-                                            height: 13
-                                            draw_icon +: { svg: (ICON_PEOPLE), color: (RBX_FG_SECONDARY) }
-                                            icon_walk: Walk{width: 13, height: 13}
-                                        }
-                                        members_meta_value := Label {
-                                            width: Fit
-                                            height: Fit
-                                            draw_text +: { text_style: RBX_TEXT_META {}, color: (RBX_FG_SECONDARY) }
-                                            text: ""
-                                        }
-                                    }
-
-                                    encryption_meta := View {
-                                        width: Fit
-                                        height: Fit
-                                        flow: Right
-                                        spacing: 4
-                                        align: Align{y: 0.5}
-
-                                        enc_icon_locked := View {
-                                            width: Fit
-                                            height: Fit
-                                            Icon {
-                                                width: 13
-                                                height: 13
-                                                draw_icon +: { svg: (ICON_LOCK), color: (RBX_SUCCESS_FG) }
-                                                icon_walk: Walk{width: 13, height: 13}
-                                            }
-                                        }
-                                        enc_icon_unlocked := View {
-                                            visible: false
-                                            width: Fit
-                                            height: Fit
-                                            Icon {
-                                                width: 13
-                                                height: 13
-                                                draw_icon +: { svg: (ICON_LOCK_OPEN), color: (RBX_FG_TERTIARY) }
-                                                icon_walk: Walk{width: 13, height: 13}
-                                            }
-                                        }
-                                        encryption_value := Label {
-                                            width: Fit
-                                            height: Fit
-                                            draw_text +: { text_style: RBX_TEXT_META {}, color: (RBX_FG_SECONDARY) }
-                                            text: ""
-                                        }
-                                    }
-                                }
-
-                                badges_row := View {
-                                    width: Fill
-                                    height: Fit
-                                    flow: Flow.Right{wrap: true}
-                                    spacing: 8
-
-                                    agent_badge_wrap := View {
-                                        visible: false
-                                        width: Fit
-                                        height: Fit
-                                        // Outdent by the pill's left padding so the
-                                        // robot icon lines up with the meta-row icons
-                                        // (globe / people) directly above.
-                                        margin: Inset{left: -10}
-
-                                        agent_badge := RoundedView {
-                                            width: Fit
-                                            height: Fit
-                                            flow: Right
-                                            spacing: 5
-                                            align: Align{y: 0.5}
-                                            padding: Inset{left: 10, right: 11, top: 5, bottom: 5}
-                                            show_bg: true
-                                            draw_bg +: {
-                                                color: (RBX_ACCENT_SOFT)
-                                                border_radius: (RBX_RADIUS_PILL)
-                                            }
-
-                                            Icon {
-                                                width: 14
-                                                height: 14
-                                                draw_icon +: { svg: (ICON_ROBOT), color: (RBX_ACCENT) }
-                                                icon_walk: Walk{width: 14, height: 14}
-                                            }
-                                            Label {
-                                                width: Fit
-                                                height: Fit
-                                                draw_text +: { text_style: RBX_TEXT_BADGE {}, color: (RBX_ACCENT) }
-                                                text: "Agent-enabled"
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    // ===== About / topic card =====
-                    about_card := RoundedView {
-                        width: Fill
-                        height: Fit
-                        flow: Right
-                        spacing: 12
-                        align: Align{y: 0.0}
-                        padding: Inset{left: 14, right: 14, top: 13, bottom: 13}
-
-                        show_bg: true
-                        draw_bg +: {
-                            color: (RBX_BG_SURFACE)
-                            border_radius: (RBX_RADIUS_MD)
-                            border_size: 1.0
-                            border_color: (RBX_STROKE_SOFT)
-                        }
-
-                        about_icon_circle := CircleView {
-                            width: 38
-                            height: 38
-                            align: Align{x: 0.5, y: 0.5}
-                            show_bg: true
-                            draw_bg +: { color: (RBX_ACCENT_SOFT) }
-                            Icon {
-                                width: 19
-                                height: 19
-                                draw_icon +: { svg: (ICON_FILE), color: (RBX_ACCENT) }
-                                icon_walk: Walk{width: 19, height: 19}
-                            }
-                        }
-
-                        about_col := View {
-                            width: Fill
-                            height: Fit
-                            flow: Down
-                            spacing: 2
-
-                            about_title := Label {
-                                width: Fill
-                                height: Fit
-                                draw_text +: { text_style: RBX_TEXT_CARD_TITLE {}, color: (RBX_FG_PRIMARY) }
-                                text: "About"
-                            }
-
-                            topic_value := Label {
-                                width: Fill
-                                height: Fit
-                                flow: Flow.Right{wrap: true}
-                                draw_text +: { text_style: RBX_TEXT_META {}, color: (RBX_FG_SECONDARY) }
-                                text: ""
-                            }
-
-                            topic_toggle_button := RobrixNeutralIconButton {
-                                visible: false
-                                width: Fit
-                                height: 24
-                                align: Align{x: 0.0, y: 0.5}
-                                padding: Inset{left: 0, right: 0, top: 3, bottom: 0}
-                                spacing: 0
-                                icon_walk: Walk{width: 0, height: 0}
-                                draw_bg +: {
-                                    color: #00000000
-                                    color_hover: #00000000
-                                    color_down: #00000000
-                                    border_size: 0.0
-                                }
-                                draw_text +: { color: (RBX_ACCENT), text_style: RBX_TEXT_META {} }
-                                text: "Expand"
-                            }
-                        }
-                    }
-
-                    // ===== Members card =====
-                    members_card := RoundedView {
-                        width: Fill
-                        height: Fit
-                        flow: Down
-                        spacing: 7
-                        padding: Inset{left: 14, right: 14, top: 13, bottom: 13}
-
-                        show_bg: true
-                        draw_bg +: {
-                            color: (RBX_BG_SURFACE)
-                            border_radius: (RBX_RADIUS_MD)
-                            border_size: 1.0
-                            border_color: (RBX_STROKE_SOFT)
-                        }
-
-                        members_header := View {
-                            width: Fill
-                            height: Fit
-                            flow: Right
-                            spacing: 12
-                            align: Align{y: 0.5}
-                            cursor: MouseCursor.Hand
-
-                            members_icon_circle := CircleView {
-                                width: 38
-                                height: 38
-                                align: Align{x: 0.5, y: 0.5}
-                                show_bg: true
-                                draw_bg +: { color: (RBX_INFO_BG) }
-                                Icon {
-                                    width: 19
-                                    height: 19
-                                    draw_icon +: { svg: (ICON_PEOPLE), color: (RBX_INFO_FG) }
-                                    icon_walk: Walk{width: 19, height: 19}
-                                }
-                            }
-
-                            members_title := Label {
-                                width: Fit
-                                height: Fit
-                                draw_text +: { text_style: RBX_TEXT_CARD_TITLE {}, color: (RBX_FG_PRIMARY) }
-                                text: "Members"
-                            }
-
-                            members_header_spacer := View { width: Fill, height: Fit }
-
-                            // count + chevron kept close together on the right edge
-                            members_count_group := View {
-                                width: Fit
-                                height: Fit
-                                flow: Right
-                                spacing: 2
-                                align: Align{y: 0.5}
-
-                                members_count_value := Label {
-                                    width: Fit
-                                    height: Fit
-                                    draw_text +: { text_style: RBX_TEXT_BODY_STRONG {}, color: (RBX_FG_SECONDARY) }
-                                    text: ""
-                                }
-
-                                Icon {
-                                    width: 14
-                                    height: 14
-                                    draw_icon +: { svg: (ICON_CHEVRON_RIGHT), color: (RBX_FG_TERTIARY) }
-                                    icon_walk: Walk{width: 14, height: 14}
-                                }
-                            }
-                        }
-
-                        members_detail := View {
-                            width: Fill
-                            height: Fit
-                            flow: Right
-                            align: Align{y: 0.5}
-                            spacing: 8
-                            // Indent so the avatar row lines up under "Members"
-                            // (icon width 38 + header spacing 12).
-                            margin: Inset{left: 50}
-
-                            members_stack := View {
-                                width: Fit
-                                height: Fit
-                                flow: Right
-                                align: Align{y: 0.5}
-
-                                stack_slot_0 := View {
-                                    width: Fit
-                                    height: Fit
-                                    ring_0 := RoundedView {
-                                        width: Fit
-                                        height: Fit
-                                        padding: 2
-                                        show_bg: true
-                                        draw_bg +: { color: (RBX_BG_SURFACE), border_radius: 18.0 }
-                                        stack_avatar_0 := Avatar { width: 32, height: 32 }
-                                    }
-                                }
-                                stack_slot_1 := View {
-                                    visible: false
-                                    width: Fit
-                                    height: Fit
-                                    margin: Inset{left: -11}
-                                    ring_1 := RoundedView {
-                                        width: Fit
-                                        height: Fit
-                                        padding: 2
-                                        show_bg: true
-                                        draw_bg +: { color: (RBX_BG_SURFACE), border_radius: 18.0 }
-                                        stack_avatar_1 := Avatar { width: 32, height: 32 }
-                                    }
-                                }
-                                stack_slot_2 := View {
-                                    visible: false
-                                    width: Fit
-                                    height: Fit
-                                    margin: Inset{left: -11}
-                                    ring_2 := RoundedView {
-                                        width: Fit
-                                        height: Fit
-                                        padding: 2
-                                        show_bg: true
-                                        draw_bg +: { color: (RBX_BG_SURFACE), border_radius: 18.0 }
-                                        stack_avatar_2 := Avatar { width: 32, height: 32 }
-                                    }
-                                }
-                                stack_more_wrap := View {
-                                    visible: false
-                                    width: Fit
-                                    height: Fit
-                                    margin: Inset{left: 1}
-                                    stack_more_chip := RoundedView {
-                                        width: Fit
-                                        height: Fit
-                                        align: Align{x: 0.5, y: 0.5}
-                                        padding: Inset{left: 7, right: 7, top: 5, bottom: 5}
-                                        show_bg: true
-                                        draw_bg +: { color: (RBX_NEUTRAL_BG), border_radius: (RBX_RADIUS_PILL) }
-                                        stack_more := Label {
-                                            width: Fit
-                                            height: Fit
-                                            draw_text +: { text_style: RBX_TEXT_BADGE {}, color: (RBX_FG_SECONDARY) }
-                                            text: ""
-                                        }
-                                    }
-                                }
-                            }
-
-                            members_detail_spacer := View { width: Fill, height: Fit }
-
-                            my_role_wrap := View {
-                                visible: false
-                                width: Fit
-                                height: Fit
-
-                                my_role_chip := RoundedView {
-                                    width: Fit
-                                    height: Fit
-                                    align: Align{x: 0.5, y: 0.5}
-                                    padding: Inset{left: 10, right: 10, top: 4, bottom: 4}
-                                    show_bg: true
-                                    draw_bg +: {
-                                        color: (RBX_ACCENT_SOFT)
-                                        border_radius: (RBX_RADIUS_PILL)
-                                    }
-                                    my_role_label := Label {
-                                        width: Fit
-                                        height: Fit
-                                        draw_text +: { text_style: RBX_TEXT_BADGE {}, color: (RBX_ACCENT) }
-                                        text: ""
-                                    }
-                                }
-                            }
-                        }
-                    }
-
-                    // ===== Actions =====
-                    actions_row := View {
-                        width: Fill
-                        height: Fit
-                        flow: Down
-                        spacing: 8
-
-                        // Primary actions grouped into ONE card (rows + hairline
-                        // divider) so they read as a settings list, not three
-                        // floating boxes. Rows are transparent; the card paints
-                        // the single border + radius.
-                        actions_card := RoundedView {
-                            width: Fill
-                            height: Fit
-                            flow: Down
-                            clip_x: true
-                            clip_y: true
-                            show_bg: true
-                            draw_bg +: {
-                                color: (RBX_BG_SURFACE)
-                                border_radius: (RBX_RADIUS_MD)
-                                border_size: 1.0
-                                border_color: (RBX_STROKE_SOFT)
-                            }
-
-                            invite_button := RobrixNeutralIconButton {
-                                width: Fill
-                                height: 50
-                                align: Align{x: 0.0, y: 0.5}
-                                spacing: 13
-                                padding: Inset{left: 16, right: 14, top: 0, bottom: 0}
-                                draw_bg +: {
-                                    color: #00000000
-                                    color_hover: (RBX_BG_HOVER)
-                                    color_down: (RBX_BG_PRESSED)
-                                    border_size: 0.0
-                                    border_radius: 0.0
-                                    border_color: #0000
-                                    border_color_hover: #0000
-                                    border_color_down: #0000
-                                }
-                                draw_icon +: { svg: (ICON_ADD_USER), color: (RBX_FG_SECONDARY) }
-                                icon_walk: Walk{width: 20, height: 20}
-                                draw_text +: {
-                                    color: (RBX_FG_PRIMARY)
-                                    color_hover: (RBX_FG_PRIMARY)
-                                    color_down: (RBX_FG_PRIMARY)
-                                    text_style: RBX_TEXT_BODY {}
-                                }
-                                text: "Invite"
-                            }
-
-                            action_divider := RoundedView {
-                                width: Fill
-                                height: 1.0
-                                margin: Inset{left: 49}
-                                show_bg: true
-                                draw_bg +: { color: (RBX_STROKE_SOFT) }
-                            }
-
-                            report_room_button := RobrixNeutralIconButton {
-                                width: Fill
-                                height: 50
-                                align: Align{x: 0.0, y: 0.5}
-                                spacing: 13
-                                padding: Inset{left: 16, right: 14, top: 0, bottom: 0}
-                                draw_bg +: {
-                                    color: #00000000
-                                    color_hover: (RBX_BG_HOVER)
-                                    color_down: (RBX_BG_PRESSED)
-                                    border_size: 0.0
-                                    border_radius: 0.0
-                                    border_color: #0000
-                                    border_color_hover: #0000
-                                    border_color_down: #0000
-                                }
-                                draw_icon +: { svg: (ICON_INFO), color: (RBX_FG_SECONDARY) }
-                                icon_walk: Walk{width: 20, height: 20}
-                                draw_text +: {
-                                    color: (RBX_FG_PRIMARY)
-                                    color_hover: (RBX_FG_PRIMARY)
-                                    color_down: (RBX_FG_PRIMARY)
-                                    text_style: RBX_TEXT_BODY {}
-                                }
-                                text: "Report room"
-                            }
-                        }
-
-                        // Destructive action — its own card, separated by the gap,
-                        // red icon + label.
-                        leave_card := RoundedView {
-                            width: Fill
-                            height: Fit
-                            flow: Down
-                            clip_x: true
-                            clip_y: true
-                            show_bg: true
-                            draw_bg +: {
-                                color: (RBX_BG_SURFACE)
-                                border_radius: (RBX_RADIUS_MD)
-                                border_size: 1.0
-                                border_color: (RBX_STROKE_SOFT)
-                            }
-
-                            leave_room_button := RobrixNegativeIconButton {
-                                width: Fill
-                                height: 50
-                                align: Align{x: 0.0, y: 0.5}
-                                spacing: 13
-                                padding: Inset{left: 16, right: 14, top: 0, bottom: 0}
-                                draw_bg +: {
-                                    color: #00000000
-                                    color_hover: (RBX_DANGER_BG)
-                                    color_down: (RBX_DANGER_BG)
-                                    border_size: 0.0
-                                    border_radius: 0.0
-                                    border_color: #0000
-                                    border_color_hover: #0000
-                                    border_color_down: #0000
-                                }
-                                draw_icon +: { svg: (ICON_CLOSE), color: (RBX_DANGER_FG) }
-                                icon_walk: Walk{width: 18, height: 18}
-                                draw_text +: {
-                                    color: (RBX_DANGER_FG)
-                                    color_hover: (RBX_DANGER_FG)
-                                    color_down: (RBX_DANGER_FG)
-                                    text_style: RBX_TEXT_BODY {}
-                                }
-                                text: "Leave Room"
-                            }
-                        }
-                    }
-                }
-
-            }
-
-            people_view := View {
-                visible: false
-                width: Fill
-                height: Fill
-                flow: Down
-                spacing: 10
-                padding: Inset{left: 12, right: 12, top: 12, bottom: 12}
-
-                loading_label := Label {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    draw_text +: {
-                        text_style: RBX_TEXT_BODY {}
-                        color: (RBX_FG_SECONDARY)
-                    }
-                    text: "Loading members..."
-                }
-
-                empty_label := Label {
-                    visible: false
-                    width: Fill
-                    height: Fit
-                    draw_text +: {
-                        text_style: RBX_TEXT_BODY {}
-                        color: (RBX_FG_SECONDARY)
-                    }
-                    text: "No members found."
-                }
-
-                // The member rows grouped into one card (rows + hairline dividers).
-                members_list_card := RoundedView {
-                    width: Fill
-                    height: Fill
-                    flow: Down
-                    clip_x: true
-                    clip_y: true
-                    show_bg: true
-                    draw_bg +: {
-                        color: (RBX_BG_SURFACE)
-                        border_radius: (RBX_RADIUS_MD)
-                        border_size: 1.0
-                        border_color: (RBX_STROKE_SOFT)
-                    }
-
-                    people_list := PortalList {
-                        width: Fill
-                        height: Fill
-                        flow: Down
-                        max_pull_down: 0.0
-
-                        PersonEntry := mod.widgets.RoomInfoPeopleEntry {}
-                    }
-                }
-            }
-        }
-
-        slide: 1.0,
-
-        animator: Animator {
-            panel: {
-                default: @hide
-                show: AnimatorState{
-                    redraw: true,
-                    from: {all: Forward {duration: 0.5}}
-                    ease: Ease.ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {
-                        slide: 0.0
-                    }
-                }
-                hide: AnimatorState{
-                    redraw: true,
-                    from: {all: Forward {duration: 0.5}}
-                    ease: Ease.ExpDecay {d1: 0.80, d2: 0.97}
-                    apply: {
-                        slide: 1.0
-                    }
-                }
-            }
-        }
-    }
-
-    mod.widgets.ReportRoomModalLabel = Label {
-        width: Fill
-        height: Fit
-        draw_text +: {
-            text_style: REGULAR_TEXT { font_size: 10.5 }
-            color: #333
-        }
-        text: ""
-    }
-
-    mod.widgets.ReportRoomModal = #(ReportRoomModal::register_widget(vm)) {
-        width: Fill { max: 430 }
-        height: Fit
-        margin: Inset{left: 12, right: 12}
-
-        RoundedShadowView {
-            width: Fill
-            height: Fit
-            align: Align{x: 0.5}
-            flow: Down
-            padding: Inset{top: 26, right: 22, bottom: 18, left: 22}
-            spacing: 14
-
-            show_bg: true
-            draw_bg +: {
-                color: (RBX_BG_SURFACE)
-                border_radius: (RBX_RADIUS_SM)
-                border_size: 1.0
-                border_color: (RBX_STROKE_SOFT)
-                shadow_color: (RBX_SHADOW_STRONG)
-                shadow_radius: 10.0
-                shadow_offset: vec2(0.0, 3.0)
-            }
-
-            title := Label {
-                width: Fill
-                height: Fit
-                draw_text +: {
-                    text_style: TITLE_TEXT { font_size: 13 }
-                    color: (RBX_FG_PRIMARY)
-                }
-                text: "Report Room"
-            }
-
-            body := mod.widgets.ReportRoomModalLabel {
-                text: ""
-            }
-
-            reason_input := RobrixTextInput {
-                width: Fill
-                height: Fit
-                padding: 10
-                draw_text +: {
-                    text_style: REGULAR_TEXT { font_size: 11.5 }
-                    color: #000
-                }
-                empty_text: "Describe why you are reporting this room"
-            }
-
-            status_label := Label {
-                width: Fill
-                height: Fit
-                draw_text +: {
-                    text_style: REGULAR_TEXT { font_size: 10.2 }
-                    color: #000
-                }
-                text: ""
-            }
-
-            buttons := View {
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{x: 1.0, y: 0.5}
-                spacing: 16
-
-                cancel_button := RobrixNeutralIconButton {
-                    width: 110
-                    align: Align{x: 0.5, y: 0.5}
-                    padding: 12
-                    draw_icon.svg: (ICON_FORBIDDEN)
-                    icon_walk: Walk{width: 16, height: 16, margin: Inset{left: -2, right: -1}}
-                    text: "Cancel"
-                }
-
-                report_button := RobrixNegativeIconButton {
-                    width: 130
-                    align: Align{x: 0.5, y: 0.5}
-                    padding: 12
-                    draw_icon.svg: (ICON_INFO)
-                    icon_walk: Walk{width: 16, height: 16, margin: Inset{left: -2, right: -1}}
-                    text: "Report room"
-                }
-            }
-        }
-    }
-
-    mod.widgets.AppServicePanel = #(AppServicePanel::register_widget(vm)) {
-        width: Fill
-        height: Fit
-        margin: Inset{left: 14, right: 54, top: 10, bottom: 16}
-        flow: Down
-        align: Align{x: 0.0, y: 0.0}
-        spacing: 8
-
-        sender_row := View {
-            width: Fit
-            height: Fit
-            flow: Right
-            spacing: 6
-
-            sender_name := Label {
-                width: Fit
-                height: Fit
-                draw_text +: {
-                    text_style: USERNAME_TEXT_STYLE { font_size: 10.8 }
-                    color: (COLOR_ACTIVE_PRIMARY)
-                }
-                text: ""
-            }
-
-            sender_tag := Label {
-                width: Fit
-                height: Fit
-                draw_text +: {
-                    text_style: REGULAR_TEXT { font_size: 9.5 }
-                    color: #8A8A8A
-                }
-                text: ""
-            }
-        }
-
-        bubble := RoundedView {
-            width: 408
-            height: Fit
-            flow: Down
-            spacing: 8
-            padding: Inset{top: 14, right: 14, bottom: 12, left: 14}
-
-            show_bg: true
-            draw_bg +: {
-                color: (COLOR_PRIMARY)
-                border_radius: 0.0
-                border_size: 1.0
-                border_color: (COLOR_SECONDARY_DARKER)
-            }
-
-            header := View {
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{y: 0.5}
-
-                title := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: {
-                        text_style: USERNAME_TEXT_STYLE { font_size: 11.2 }
-                        color: #1F1F1F
-                    }
-                    text: ""
-                }
-
-                spacer := View {
-                    width: Fill
-                    height: Fit
-                }
-
-                dismiss_button := RobrixNeutralIconButton {
-                    width: 28
-                    height: 24
-                    align: Align{x: 0.5, y: 0.5}
-                    spacing: 0
-                    padding: 0
-                    draw_icon.svg: (ICON_CLOSE)
-                    icon_walk: Walk{width: 12, height: 12}
-                    text: ""
-                }
-            }
-
-            subtitle := Label {
-                width: Fill
-                height: Fit
-                draw_text +: {
-                    text_style: REGULAR_TEXT { font_size: 10.5 }
-                    color: (COLOR_TEXT)
-                }
-                text: ""
-            }
-
-            footer := View {
-                width: Fill
-                height: Fit
-                flow: Right
-                align: Align{x: 1.0, y: 0.5}
-
-                timestamp := Label {
-                    width: Fit
-                    height: Fit
-                    draw_text +: {
-                        text_style: REGULAR_TEXT { font_size: 8.8 }
-                        color: #9A9A9A
-                    }
-                    text: ""
-                }
-            }
-        }
-
-        keyboard := View {
-            width: Fit
-            height: Fit
-            flow: Down
-            spacing: 8
-
-            first_row := View {
-                width: Fit
-                height: Fit
-                flow: Right
-                spacing: 8
-
-                create_button := RobrixPositiveIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_CHECKMARK)
-                    icon_walk: Walk{width: 16, height: 16, margin: Inset{left: -2, right: -1}}
-                    text: ""
-                }
-
-                list_button := RobrixNeutralIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_SEARCH)
-                    icon_walk: Walk{width: 14, height: 14, margin: Inset{left: -2, right: -1}}
-                    text: ""
-                }
-            }
-
-            second_row := View {
-                width: Fit
-                height: Fit
-                flow: Right
-                spacing: 8
-
-                delete_button := RobrixNegativeIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_CLOSE)
-                    icon_walk: Walk{width: 14, height: 14, margin: Inset{left: -2, right: -1}}
-                    text: ""
-                }
-
-                help_button := RobrixNeutralIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_INFO)
-                    icon_walk: Walk{width: 14, height: 14, margin: Inset{left: -2, right: -1}}
-                    text: ""
-                }
-            }
-
-            third_row := View {
-                width: Fit
-                height: Fit
-                flow: Right
-                spacing: 8
-
-                view_bound_button := RobrixNeutralIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_SEARCH)
-                    icon_walk: Walk{width: 14, height: 14, margin: Inset{left: -2, right: -1}}
-                    text: "View Bound Bots"
-                }
-
-                unbind_button := RobrixNeutralIconButton {
-                    width: 156
-                    height: 46
-                    padding: 10
-                    draw_icon.svg: (ICON_CLOSE)
-                    icon_walk: Walk{width: 14, height: 14, margin: Inset{left: -2, right: -1}}
-                    text: ""
-                }
-            }
-        }
-    }
-
-    mod.widgets.Timeline = View {
-        width: Fill,
-        height: Fill,
-        align: Align{x: 0.5, y: 0.0} // center horizontally, align to top vertically
-        flow: Overlay,
-
-        list := PortalList {
-            height: Fill,
-            width: Fill
-            flow: Down
-
-            auto_tail: true, // set to `true` to lock the view to the last item.
-            max_pull_down: 0.0, // set to `0.0` to disable the pulldown bounce animation.
-            // TODO: enable `reuse_items: true` once Makepad's Html/TextFlow widget
-            //   properly resets all internal state during `script_apply(Reload)`.
-            //   Currently, stale TextFlow layout state (particularly related to
-            //   list items) leaks through when a widget is recycled, causing
-            //   excessive whitespace in HTML messages with `<ul>`/`<ol>` lists.
-
-            // Below, we must place all of the possible templates (views) that can be used in the portal list.
-            Message := mod.widgets.Message {}
-            CondensedMessage := mod.widgets.CondensedMessage {}
-            ImageMessage := mod.widgets.ImageMessage {}
-            CondensedImageMessage := mod.widgets.CondensedImageMessage {}
-            VideoMessage := mod.widgets.VideoMessage {}
-            AudioMessage := mod.widgets.AudioMessage {}
-            StickerMessage := mod.widgets.StickerMessage {}
-            CondensedStickerMessage := mod.widgets.CondensedStickerMessage {}
-            SmallStateEvent := mod.widgets.SmallStateEvent {}
-            SmallStateEventsSummary := mod.widgets.SmallStateEventsSummary {}
-            Empty := mod.widgets.Empty {}
-            EncryptionNotice := mod.widgets.EncryptionNotice {}
-            DateDivider := mod.widgets.DateDivider {}
-            ReadMarker := mod.widgets.ReadMarker {}
-            AppServicePanel := mod.widgets.AppServicePanel {}
-        }
-
-        // A jump to bottom button (with an unread message badge) that is shown
-        // when the timeline is not at the bottom.
-        jump_to_bottom_button := JumpToBottomButton { }
-
-        // Floating info button at the top-right, occupying the rightmost slot.
-        // Clicking it opens the `room_info_sliding_pane` (desktop only).
-        info_button := mod.widgets.InfoButton { }
-
-        // Floating threads button at the top-right, sitting left of the search
-        // and info buttons. Clicking it opens the `threads_sliding_pane`.
-        threads_button := mod.widgets.ThreadsButton { }
-
-        // Floating search button at the top-right (mirrors jump-to-bottom).
-        // Clicking it opens the `search_messages_pane` sliding pane.
-        // NOTE: the pane itself is NOT defined here — it lives as a top-level
-        // overlay in `room_screen_wrapper` (next to `room_info_sliding_pane`)
-        // so it composites OVER the timeline's `new_batch` cards. Defining it
-        // inside this Timeline overlay let those cards z-fight through it.
-        search_messages_button := mod.widgets.SearchMessagesButton { }
-    }
-
-    mod.widgets.TranslationLangPopupButton = RobrixIconButton {
-        width: Fill
-        height: 36
-        spacing: 0
-        margin: 0
-        padding: Inset{left: 12, right: 12, top: 8, bottom: 8}
-        icon_walk: Walk{width: 0, height: 0}
-        draw_text +: {
-            color: (COLOR_TEXT)
-            color_hover: (COLOR_TEXT)
-            color_down: (COLOR_TEXT)
-            text_style: MESSAGE_TEXT_STYLE { font_size: 10.5 }
-        }
-        draw_bg +: {
-            color: #0000
-            color_hover: #xF0F4FA
-            color_down: #xE8EEF8
-            border_size: 0.0
-            border_radius: 0.0
-        }
-    }
-
-    mod.widgets.RoomScreen = #(RoomScreen::register_widget(vm)) {
-        width: Fill, height: Fill,
-        cursor: MouseCursor.Default,
-        flow: Down,
-        spacing: 0.0
-
-        room_screen_wrapper := SolidView {
-            width: Fill, height: Fill,
-            flow: Overlay,
-
-            show_bg: true
-            draw_bg.color: (COLOR_PRIMARY_DARKER)
-
-            restore_status_view := RestoreStatusView {}
-
-            // Keyboard avoidance is already provided by the Window's built-in
-            // KeyboardView (makepad `window.rs`: `body := KeyboardView`). Using a
-            // second KeyboardView here nested inside it double-applies the
-            // keyboard shift (content jumps / a big blank gap on Android), so
-            // this is a plain View.
-            keyboard_view := View {
-                width: Fill, height: Fill,
-                flow: Down,
-
-                // Robrix-owned mobile room header + `Chat | Info` tab row.
-                // Hidden on desktop (which uses its own dock chrome); shown and
-                // populated on mobile from `draw_walk`.
-                room_top_bar := mod.widgets.RoomTopBar {
-                    visible: false,
-                }
-
-                // The Chat / Info bodies share the same space via an Overlay so
-                // exactly one is shown at a time. (Two `height: Fill` siblings
-                // in a Down flow are sized incorrectly when one is hidden — the
-                // hidden one's Fill space can still be reserved, pushing the
-                // visible one off-screen. An Overlay sizes BOTH children to the
-                // full area, so toggling visibility just swaps which is drawn.)
-                body_area := View {
-                    width: Fill, height: Fill,
-                    flow: Overlay,
-
-                    // "Chat" tab body: the timeline, typing notice, and input bar.
-                    chat_content := View {
-                        width: Fill, height: Fill,
-                        flow: Down,
-
-                        // First, display the timeline of all messages/events.
-                        timeline := mod.widgets.Timeline {
-                            // margin: Inset{bottom: 10}
-                        }
-
-                        // Below that, display a typing notice when other users in the room are typing.
-                        typing_notice := TypingNotice { }
-
-                        room_input_bar := RoomInputBar {
-                            // margin: Inset{top: 20}
-                        }
-                    }
-
-                    // "Info" tab body: the existing room-info content reused
-                    // inline (no slide animation / backdrop / close button).
-                    //
-                    // NOTE: the visibility toggle is on this PLAIN `View`
-                    // wrapper — not on `info_content` directly — because
-                    // `set_visible` is a no-op on a custom widget
-                    // (`RoomInfoSlidingPane`); only plain Views honor it. The
-                    // inner pane stays always-visible + inline.
-                    info_tab_body := View {
-                        width: Fill, height: Fill,
-                        visible: false,
-
-                        info_content := mod.widgets.RoomInfoSlidingPane {
-                            // The base RoomInfoSlidingPane is `visible: false`
-                            // (it's normally a hidden slide-in pane); override
-                            // to always-visible here since the `info_tab_body`
-                            // wrapper controls when this inline copy is shown.
-                            visible: true,
-                            inline: true,
-                            bg_view +: { visible: false }
-                            main_content +: {
-                                width: Fill
-                                header +: {
-                                    padding: 0
-                                    title +: { visible: false }
-                                    close_button +: { visible: false }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-
-            translation_lang_modal := Modal {
-                align: Align{x: 0, y: 0}
-                bg_view.draw_bg.color: #00000000
-                content +: {
-                    width: Fill
-                    height: Fill
-                    flow: Overlay
-                    align: Align{x: 0, y: 0}
-
-                    translation_lang_popup := RoundedView {
-                        width: 220
-                        height: Fit
-                        margin: Inset{left: 0, top: 0}
-                        padding: Inset{top: 4, bottom: 4}
-                        show_bg: true
-                        new_batch: true
-                        draw_bg +: {
-                            color: (COLOR_PRIMARY)
-                            border_radius: 6.0
-                            border_size: 1.0
-                            border_color: #ddd
-                            shadow_color: #0003
-                            shadow_radius: 8.0
-                            shadow_offset: vec2(0.0, 2.0)
-                        }
-
-                        translation_lang_scroll := ScrollYView {
-                            width: Fill
-                            height: 288
-                            flow: Down
-                            spacing: 0
-
-                            lang_en := mod.widgets.TranslationLangPopupButton { text: "en  English" }
-                            lang_zh := mod.widgets.TranslationLangPopupButton { text: "zh  简体中文" }
-                            lang_zh_tw := mod.widgets.TranslationLangPopupButton { text: "zh-TW  繁體中文" }
-                            lang_ja := mod.widgets.TranslationLangPopupButton { text: "ja  日本語" }
-                            lang_ko := mod.widgets.TranslationLangPopupButton { text: "ko  한국어" }
-                            lang_es := mod.widgets.TranslationLangPopupButton { text: "es  Español" }
-                            lang_fr := mod.widgets.TranslationLangPopupButton { text: "fr  Français" }
-                            lang_de := mod.widgets.TranslationLangPopupButton { text: "de  Deutsch" }
-                            lang_ru := mod.widgets.TranslationLangPopupButton { text: "ru  Русский" }
-                            lang_pt := mod.widgets.TranslationLangPopupButton { text: "pt  Português" }
-                            lang_ar := mod.widgets.TranslationLangPopupButton { text: "ar  العربية" }
-                            lang_vi := mod.widgets.TranslationLangPopupButton { text: "vi  Tiếng Việt" }
-                            lang_th := mod.widgets.TranslationLangPopupButton { text: "th  ไทย" }
-                            lang_id := mod.widgets.TranslationLangPopupButton { text: "id  Bahasa Indonesia" }
-                            lang_ms := mod.widgets.TranslationLangPopupButton { text: "ms  Bahasa Melayu" }
-                            lang_tr := mod.widgets.TranslationLangPopupButton { text: "tr  Türkçe" }
-                            lang_hi := mod.widgets.TranslationLangPopupButton { text: "hi  हिन्दी" }
-                        }
-                    }
-                }
-            }
-
-            // Note: here, we're within a View that has an Overlay flow,
-            // so the order that we define the below views determines which one is on top.
-
-            // The top space should be displayed as an overlay at the top of the timeline.
-            top_space := mod.widgets.TopSpace { }
-
-            threads_sliding_pane := mod.widgets.ThreadsSlidingPane { }
-            room_info_sliding_pane := mod.widgets.RoomInfoSlidingPane { }
-            // Right-sliding pane hosting the server-side message search. Lives
-            // here (a top-level wrapper overlay), NOT inside the Timeline, so it
-            // composites over the timeline's `new_batch` message cards — same
-            // as `room_info_sliding_pane`, which has no z-order glitch.
-            search_messages_pane := mod.widgets.SearchMessagesSlidingPane { }
-
-            // The user profile sliding pane should be displayed on top of other "static" subviews
-            // (on top of all other views that are always visible).
-            user_profile_sliding_pane := mod.widgets.UserProfileSlidingPane { }
-
-            // The loading pane appears while the user is waiting for something in the room screen
-            // to finish loading, e.g., when loading an older replied-to message.
-            loading_pane := LoadingPane { }
-
-            create_bot_modal := Modal {
-                content +: {
-                    create_bot_modal_inner := mod.widgets.CreateBotModal {}
-                }
-            }
-
-            delete_bot_modal := Modal {
-                content +: {
-                    delete_bot_modal_inner := mod.widgets.DeleteBotModal {}
-                }
-            }
-
-            /*
-             * TODO: add the action bar back in as a series of floating buttons.
-             *
-            message_action_bar_popup := PopupNotification {
-                align: Align{x: 0.0, y: 0.0}
-                content: {
-                    height: Fit,
-                    width: Fit,
-                    show_bg: false,
-                    align: Align{
-                        x: 0.5,
-                        y: 0.5
-                    }
-
-                    message_action_bar := MessageActionBar {}
-                }
-            }
-            */
-        }
-    }
-}
-
-#[derive(Clone, Default, Debug)]
-pub enum ThreadsPaneAction {
-    OpenThread(OwnedEventId),
-    LoadMoreRequested,
-    /// The pane's close button (or Esc / back / click-outside) was triggered.
-    /// The room screen should call `hide_threads_pane` to animate the pane
-    /// out and re-show the floating threads button.
-    CloseRequested,
-    #[default]
-    None,
-}
-
-impl ActionDefaultRef for ThreadsPaneAction {
-    fn default_ref() -> &'static Self {
-        static DEFAULT: ThreadsPaneAction = ThreadsPaneAction::None;
-        &DEFAULT
-    }
-}
-
-#[derive(Clone, Default, Debug)]
-pub enum ThreadsButtonAction {
-    OpenRequested,
-    #[default]
-    None,
-}
-
-impl ActionDefaultRef for ThreadsButtonAction {
-    fn default_ref() -> &'static Self {
-        static DEFAULT: ThreadsButtonAction = ThreadsButtonAction::None;
-        &DEFAULT
-    }
-}
-
-#[derive(Clone, Default, Debug)]
-pub enum InfoButtonAction {
-    OpenRequested,
-    #[default]
-    None,
-}
-
-impl ActionDefaultRef for InfoButtonAction {
-    fn default_ref() -> &'static Self {
-        static DEFAULT: InfoButtonAction = InfoButtonAction::None;
-        &DEFAULT
-    }
-}
-
-#[derive(Clone, Default, Debug)]
-pub enum RoomInfoPaneAction {
-    InviteUser,
-    ShowPeoplePage,
-    /// Emitted ONLY by a `RoomInfoPeopleEntry` row when its person is tapped.
-    /// The owning `RoomInfoSlidingPane` instance re-bubbles this as
-    /// `OpenPeopleProfile` (tagged with the pane's own widget uid) so the
-    /// `RoomScreen` handler can pick it up. Kept distinct from
-    /// `OpenPeopleProfile` so that pane instances never react to each other's
-    /// re-emitted action — otherwise the desktop overlay pane and the inline
-    /// (mobile) pane, which both receive every broadcast `Event::Actions`,
-    /// would ping-pong the action between themselves forever and freeze the app.
-    PersonClicked(OwnedUserId),
-    OpenPeopleProfile(OwnedUserId),
-    ReportRoom,
-    LeaveRoom,
-    #[default]
-    None,
-}
-
-impl ActionDefaultRef for RoomInfoPaneAction {
-    fn default_ref() -> &'static Self {
-        static DEFAULT: RoomInfoPaneAction = RoomInfoPaneAction::None;
-        &DEFAULT
-    }
-}
-
-#[derive(Clone, Debug)]
-struct ThreadsPaneEntryInfo {
-    thread_root_event_id: OwnedEventId,
-    title: String,
-    subtitle: String,
-    time: String,
-    preview: String,
-}
-
-#[derive(Clone, Debug)]
-struct ThreadsPaneInfo {
-    room_name: String,
-    entries: Vec<ThreadsPaneEntryInfo>,
-    status_text: String,
-    show_entries: bool,
-    loading_text: String,
-    show_loading: bool,
-}
-
-#[derive(Clone, Debug)]
-struct RoomInfoPaneInfo {
-    room_name: String,
-    room_id: String,
-    /// Owned room id kept alongside the display string so the info pane can
-    /// issue room-scoped requests (e.g. toggling the favourite tag).
-    owned_room_id: OwnedRoomId,
-    topic: String,
-    /// Short visibility label: "Public" / "Private" / "Unknown".
-    visibility: String,
-    /// Short encryption label: "Encrypted" / "Unencrypted" / "Unknown".
-    encryption: String,
-    is_encrypted: bool,
-    /// Whether the room carries the `m.favourite` tag for the current user.
-    is_favorite: bool,
-    /// Whether this room has a bot/agent participating (any member detected as a
-    /// bot via `is_likely_bot_member`).
-    is_agent_enabled: bool,
-    /// Whether the compact "bot" pill should trail the room title. This follows
-    /// the same user-facing badge semantics as the rooms list: room binding or a
-    /// registered-agent DM, not merely a bot-looking member in an unbound room.
-    show_title_bot_pill: bool,
-    member_count: usize,
-    /// The current user's role in this room: "Owner" / "Admin" / "Moderator" /
-    /// "Member", or empty if members haven't loaded yet.
-    my_role: String,
-    room_avatar_uri: Option<OwnedMxcUri>,
-    room_avatar_fallback_text: String,
-    /// The sorted member rows, shared via `Arc` so the (potentially huge) list is
-    /// built/sorted at most once per member-list change and cheaply reference-
-    /// counted into every subsequent Signal-driven refresh.
-    people_entries: Arc<Vec<RoomInfoPeopleEntryInfo>>,
-    people_count_text: String,
-    show_people_loading: bool,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-struct RoomInfoBotIdentityFingerprint {
-    resolved_parent_bot_user_id: Option<OwnedUserId>,
-    known_bot_user_ids: Vec<OwnedUserId>,
-}
-
-fn room_info_bot_identity_fingerprint(
-    app_state: Option<&AppState>,
-    my_user_id: Option<&UserId>,
-) -> RoomInfoBotIdentityFingerprint {
-    app_state
-        .map(|app_state| {
-            let resolved_parent_bot_user_id = if app_state.bot_settings.enabled {
-                app_state
-                    .bot_settings
-                    .resolved_bot_user_id(my_user_id)
-                    .ok()
-            } else {
-                None
-            };
-            RoomInfoBotIdentityFingerprint {
-                resolved_parent_bot_user_id,
-                known_bot_user_ids: timeline_known_bot_user_ids(app_state),
-            }
-        })
-        .unwrap_or(RoomInfoBotIdentityFingerprint {
-            resolved_parent_bot_user_id: None,
-            known_bot_user_ids: Vec::new(),
-        })
-}
-
-/// Delegates to the rooms-list predicate (`room_shows_agent_badge`) so the
-/// Info-pane title pill and the rooms-list row pill always agree.
-fn room_info_title_shows_agent_badge<'a>(
-    app_state: Option<&AppState>,
-    room_id: &RoomId,
-    dm_target: Option<&UserId>,
-    member_user_ids: impl IntoIterator<Item = &'a UserId>,
-) -> bool {
-    app_state.is_some_and(|app_state|
-        room_shows_agent_badge(app_state, room_id, dm_target, member_user_ids)
-    )
-}
-
-fn room_info_dm_target_from_user_ids<'a>(
-    user_ids: impl IntoIterator<Item = &'a UserId>,
-    my_user_id: Option<&UserId>,
-) -> Option<OwnedUserId> {
-    let mut dm_target = None;
-    for user_id in user_ids {
-        if my_user_id.is_some_and(|my_user_id| my_user_id == user_id) {
-            continue;
-        }
-        if dm_target.is_some() {
-            return None;
-        }
-        dm_target = Some(user_id.to_owned());
-    }
-    dm_target
-}
-
-/// Cache for the expensive member-row build. Keyed by the room and the identity
-/// (`Arc` pointer) of `TimelineUiState::room_members`, which is replaced wholesale
-/// whenever the member list changes — so a pointer match means "members unchanged,
-/// reuse the prebuilt rows" and we skip rebuilding + re-sorting all members on
-/// every sync Signal (critical for very large rooms). Bot identity context is part
-/// of the key because registry / app-service updates can change row bot markers
-/// without changing the room member list.
-struct RoomInfoMembersCache {
-    room_id: OwnedRoomId,
-    /// The exact `room_members` `Arc` the cached rows were built from. Held so the
-    /// allocation can't be freed and its address reused (an ABA false-hit), and so
-    /// validity is a cheap `Arc::ptr_eq` against the current `room_members`.
-    members: Arc<Vec<RoomMember>>,
-    bot_identity: RoomInfoBotIdentityFingerprint,
-    entries: Arc<Vec<RoomInfoPeopleEntryInfo>>,
-    is_agent_enabled: bool,
-    my_role: String,
-}
-
-#[derive(Clone, Debug)]
-struct RoomInfoPeopleEntryInfo {
-    user_id: OwnedUserId,
-    display_name: String,
-    level: String,
-    is_bot: bool,
-    avatar_uri: Option<OwnedMxcUri>,
-    avatar_fallback_text: String,
-}
-
-#[derive(Default)]
-struct ThreadsPaneState {
-    room_id: Option<OwnedRoomId>,
-    entries: Vec<FetchedRoomThread>,
-    prev_batch_token: Option<String>,
-    is_loading: bool,
-    initialized: bool,
-    status_text: String,
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct ThreadsPaneEntry {
-    #[source] source: ScriptObjectRef,
-    #[deref] view: View,
-
-    #[rust] thread_root_event_id: Option<OwnedEventId>,
-}
-
-impl Widget for ThreadsPaneEntry {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        // Hit-test the parent area BEFORE propagating to children, so the inner
-        // HtmlOrPlaintext (and its TextFlow / HtmlLink children) don't steal
-        // FingerDown/Up — mirrors the pattern in rooms_list_entry.rs.
-        if let Some(thread_root_event_id) = self.thread_root_event_id.clone() {
-            let area = self.view.area();
-            match event.hits(cx, area) {
-                Hit::FingerDown(_) => {
-                    cx.set_key_focus(area);
-                }
-                Hit::FingerUp(fe) if fe.is_over && fe.is_primary_hit() && fe.was_tap() => {
-                    log!("ThreadsPaneEntry: tap detected, emitting OpenThread({})", thread_root_event_id);
-                    cx.widget_action(
-                        self.widget_uid(),
-                        ThreadsPaneAction::OpenThread(thread_root_event_id),
-                    );
-                }
-                _ => {}
-            }
-        }
-
-        self.view.handle_event(cx, event, scope);
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-impl ThreadsPaneEntry {
-    fn set_entry(&mut self, cx: &mut Cx, entry: &ThreadsPaneEntryInfo) {
-        self.thread_root_event_id = Some(entry.thread_root_event_id.clone());
-        self.html_or_plaintext(cx, ids!(title)).show_html(cx, &entry.title);
-        self.label(cx, ids!(time)).set_text(cx, &entry.time);
-        self.label(cx, ids!(subtitle)).set_text(cx, &entry.subtitle);
-        self.html_or_plaintext(cx, ids!(preview)).show_html(cx, &entry.preview);
-    }
-}
-
-impl ThreadsPaneEntryRef {
-    fn set_entry(&self, cx: &mut Cx, entry: &ThreadsPaneEntryInfo) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.set_entry(cx, entry);
-    }
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct RoomInfoPeopleEntry {
-    #[source] source: ScriptObjectRef,
-    #[deref] view: View,
-
-    #[rust] user_id: Option<OwnedUserId>,
-}
-
-impl Widget for RoomInfoPeopleEntry {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        self.view.handle_event(cx, event, scope);
-
-        let Some(user_id) = self.user_id.clone() else { return };
-        match event.hits(cx, self.view.area()) {
-            Hit::FingerUp(fe) if fe.is_over && fe.is_primary_hit() && fe.was_tap() => {
-                cx.widget_action(
-                    self.widget_uid(),
-                    RoomInfoPaneAction::PersonClicked(user_id),
-                );
-            }
-            _ => {}
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-impl RoomInfoPeopleEntry {
-    fn set_entry(&mut self, cx: &mut Cx, entry: &RoomInfoPeopleEntryInfo) {
-        self.user_id = Some(entry.user_id.clone());
-        self.label(cx, ids!(display_name)).set_text(cx, &entry.display_name);
-        self.view(cx, ids!(bot_badge)).set_visible(cx, entry.is_bot);
-        self.label(cx, ids!(level)).set_text(cx, &entry.level);
-        self.view(cx, ids!(level_chip)).set_visible(cx, !entry.level.is_empty());
-
-        let avatar = self.avatar(cx, ids!(avatar));
-        if let Some(uri) = entry.avatar_uri.as_ref()
-            && let avatar_cache::AvatarCacheEntry::Loaded(image_data) = avatar_cache::get_or_fetch_avatar(cx, uri)
-        {
-            let res = avatar.show_image(
-                cx,
-                None,
-                |cx, img_ref| utils::load_png_or_jpg(&img_ref, cx, &image_data),
-            );
-            if res.is_err() {
-                avatar.show_text(cx, None, None, &entry.avatar_fallback_text);
-            }
-        } else {
-            avatar.show_text(cx, None, None, &entry.avatar_fallback_text);
-        }
-    }
-}
-
-impl RoomInfoPeopleEntryRef {
-    fn set_entry(&self, cx: &mut Cx, entry: &RoomInfoPeopleEntryInfo) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.set_entry(cx, entry);
-    }
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct ThreadsButton {
-    #[deref] view: View,
-}
-
-impl Widget for ThreadsButton {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        let button_area = self.button(cx, ids!(inner_button)).area();
-        match event.hits(cx, button_area) {
-            Hit::FingerHoverIn(_) | Hit::FingerLongPress(_) => {
-                cx.widget_action(
-                    self.widget_uid(),
-                    TooltipAction::HoverIn {
-                        text: String::from("Threads"),
-                        widget_rect: button_area.rect(cx),
-                        options: CalloutTooltipOptions {
-                            position: TooltipPosition::Left,
-                            ..Default::default()
-                        },
-                    },
-                );
-            }
-            Hit::FingerHoverOut(_) => {
-                cx.widget_action(self.widget_uid(), TooltipAction::HoverOut);
-            }
-            _ => {}
-        }
-
-        self.view.handle_event(cx, event, scope);
-
-        if let Event::Actions(actions) = event {
-            if self.button(cx, ids!(inner_button)).clicked(actions) {
-                cx.widget_action(self.widget_uid(), ThreadsButtonAction::OpenRequested);
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct InfoButton {
-    #[deref] view: View,
-}
-
-impl Widget for InfoButton {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        let button_area = self.button(cx, ids!(inner_button)).area();
-        match event.hits(cx, button_area) {
-            Hit::FingerHoverIn(_) | Hit::FingerLongPress(_) => {
-                cx.widget_action(
-                    self.widget_uid(),
-                    TooltipAction::HoverIn {
-                        text: String::from("Room Info"),
-                        widget_rect: button_area.rect(cx),
-                        options: CalloutTooltipOptions {
-                            position: TooltipPosition::Left,
-                            ..Default::default()
-                        },
-                    },
-                );
-            }
-            Hit::FingerHoverOut(_) => {
-                cx.widget_action(self.widget_uid(), TooltipAction::HoverOut);
-            }
-            _ => {}
-        }
-
-        self.view.handle_event(cx, event, scope);
-
-        if let Event::Actions(actions) = event {
-            if self.button(cx, ids!(inner_button)).clicked(actions) {
-                cx.widget_action(self.widget_uid(), InfoButtonAction::OpenRequested);
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-#[derive(Script, ScriptHook, Widget, Animator)]
-pub struct ThreadsSlidingPane {
-    #[source] source: ScriptObjectRef,
-    #[deref] view: View,
-    #[apply_default] animator: Animator,
-    #[live] slide: f32,
-
-    #[rust] info: Option<ThreadsPaneInfo>,
-    #[rust] is_animating_out: bool,
-}
-
-impl Widget for ThreadsSlidingPane {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        self.view.handle_event(cx, event, scope);
-
-        if !self.visible { return; }
-
-        let animator_action = self.animator_handle_event(cx, event);
-        if animator_action.must_redraw() {
-            self.redraw(cx);
-        }
-
-        if self.is_animating_out && !self.animator.is_track_animating(id!(panel)) {
-            self.visible = false;
-            self.is_animating_out = false;
-            cx.revert_key_focus();
-            self.view(cx, ids!(bg_view)).set_visible(cx, false);
-            self.redraw(cx);
-            return;
-        }
-
-        let area = self.view.area();
-        let close_pane = {
-            matches!(
-                event,
-                Event::Actions(actions) if self.button(cx, ids!(close_button)).clicked(actions)
-            )
-            || event.back_pressed()
-            || match event.hits_with_capture_overload(cx, area, true) {
-                Hit::KeyUp(key) => key.key_code == KeyCode::Escape,
-                Hit::FingerDown(_fde) => {
-                    cx.set_key_focus(area);
-                    false
-                }
-                Hit::FingerUp(fue) if fue.is_over => {
-                    fue.mouse_button().is_some_and(|b| b.is_back())
-                    || !self.view(cx, ids!(main_content)).area().rect(cx).contains(fue.abs)
-                }
-                _ => false,
-            }
-        };
-        if close_pane {
-            cx.widget_action(self.widget_uid(), ThreadsPaneAction::CloseRequested);
-        }
-
-        if let Event::Actions(actions) = event {
-            let threads_list = self.portal_list(cx, ids!(threads_list));
-            if threads_list.scrolled(actions)
-                && threads_list.first_id() == 0
-                && threads_list.scroll_position() >= -0.5
-            {
-                cx.widget_action(
-                    self.widget_uid(),
-                    ThreadsPaneAction::LoadMoreRequested,
-                );
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let Some(info) = self.info.as_ref() else {
-            self.visible = false;
-            return self.view.draw_walk(cx, scope, walk);
-        };
-
-        let container_width = self.view.area().rect(cx).size.x as f32;
-        let panel_width = if container_width > 1.0 && container_width < ROOM_INFO_PANE_MOBILE_BREAKPOINT {
-            container_width
-        } else {
-            ROOM_INFO_PANE_DESKTOP_WIDTH
-        };
-        let right_margin = -(self.slide * panel_width);
-        let mut main_content = self.view(cx, ids!(main_content));
-        script_apply_eval!(cx, main_content, {
-            width: #(panel_width)
-            margin.right: #(right_margin)
-        });
-        let bg_alpha = (1.0 - self.slide) * 0.733;
-        let bg_color = vec4(0.0, 0.0, 0.0, bg_alpha);
-        let mut bg_view = self.view(cx, ids!(bg_view));
-        script_apply_eval!(cx, bg_view, {
-            draw_bg +: { color: #(bg_color) }
-        });
-
-        self.label(cx, ids!(room_name)).set_text(cx, &info.room_name);
-        self.label(cx, ids!(loading_label)).set_text(cx, &info.loading_text);
-        self.view(cx, ids!(loading_indicator)).set_visible(cx, info.show_loading);
-        self.label(cx, ids!(empty_state)).set_text(cx, &info.status_text);
-        self.view(cx, ids!(empty_state)).set_visible(cx, !info.show_entries && !info.show_loading);
-        self.view(cx, ids!(threads_list)).set_visible(cx, info.show_entries);
-
-        while let Some(widget) = self.view.draw_walk(cx, scope, walk).step() {
-            let portal_list_ref = widget.as_portal_list();
-            let Some(mut list) = portal_list_ref.borrow_mut() else { continue };
-
-            list.set_item_range(cx, 0, info.entries.len());
-            while let Some(item_id) = list.next_visible_item(cx) {
-                let Some(entry) = info.entries.get(item_id) else { continue };
-                let item = list.item(cx, item_id, id!(ThreadEntry));
-                item.as_threads_pane_entry().set_entry(cx, entry);
-                item.draw_all(cx, &mut Scope::empty());
-            }
-        }
-        DrawStep::done()
-    }
-}
-
-impl ThreadsSlidingPane {
-    pub fn is_currently_shown(&self, _cx: &mut Cx) -> bool {
-        self.visible
-    }
-
-    fn set_info(&mut self, _cx: &mut Cx, info: ThreadsPaneInfo) {
-        self.info = Some(info);
-    }
-
-    pub fn show(&mut self, cx: &mut Cx) {
-        self.visible = true;
-        self.is_animating_out = false;
-        cx.set_key_focus(self.view.area());
-        self.animator_play(cx, ids!(panel.show));
-        self.view(cx, ids!(bg_view)).set_visible(cx, true);
-        self.view.button(cx, ids!(close_button)).reset_hover(cx);
-        self.redraw(cx);
-    }
-
-    pub fn hide(&mut self, cx: &mut Cx) {
-        if !self.visible {
-            return;
-        }
-        self.is_animating_out = true;
-        self.animator_play(cx, ids!(panel.hide));
-        self.redraw(cx);
-    }
-}
-
-impl ThreadsSlidingPaneRef {
-    pub fn is_currently_shown(&self, cx: &mut Cx) -> bool {
-        let Some(inner) = self.borrow() else { return false };
-        inner.is_currently_shown(cx)
-    }
-
-    fn set_info(&self, cx: &mut Cx, info: ThreadsPaneInfo) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.set_info(cx, info);
-    }
-
-    pub fn show(&self, cx: &mut Cx) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.show(cx);
-    }
-
-    pub fn hide(&self, cx: &mut Cx) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.hide(cx);
-    }
-}
-
-#[derive(Script, ScriptHook, Widget, Animator)]
-pub struct RoomInfoSlidingPane {
-    #[source] source: ScriptObjectRef,
-    #[deref] view: View,
-    #[apply_default] animator: Animator,
-    #[live] slide: f32,
-
-    /// When `true`, this pane is mounted *inline* (e.g. as the room's "Info"
-    /// body tab) rather than as a right-sliding overlay: the slide animation,
-    /// dimmed backdrop, and tap-outside/back-to-close behavior are all skipped,
-    /// and its visibility is controlled entirely by the parent (the tab).
-    #[live] inline: bool,
-
-    #[rust] info: Option<RoomInfoPaneInfo>,
-    #[rust] is_animating_out: bool,
-    #[rust] show_people_page: bool,
-    #[rust] topic_expanded: bool,
-    #[rust] people_display_count: usize,
-    /// Optimistic favourite override `(room_id, is_favorite)` set when the user
-    /// taps the star. `room.is_favourite()` keeps returning the old value until
-    /// the async tag write syncs back, so this survives the frequent
-    /// Signal-driven `set_info` rebuilds and is cleared once the room agrees.
-    #[rust] pending_favorite: Option<(OwnedRoomId, bool)>,
-}
-
-/// Populate an `Avatar` with the member/room image if it's cached, otherwise
-/// fall back to the text initials. Shared by the room hero avatar and the
-/// members-card avatar stack.
-fn show_avatar_or_text(
-    cx: &mut Cx,
-    avatar: &AvatarRef,
-    avatar_uri: Option<&OwnedMxcUri>,
-    fallback_text: &str,
-) {
-    if let Some(uri) = avatar_uri
-        && let avatar_cache::AvatarCacheEntry::Loaded(image_data) = avatar_cache::get_or_fetch_avatar(cx, uri)
-    {
-        let res = avatar.show_image(
-            cx,
-            None,
-            |cx, img_ref| utils::load_png_or_jpg(&img_ref, cx, &image_data),
-        );
-        if res.is_err() {
-            avatar.show_text(cx, None, None, fallback_text);
-        }
-    } else {
-        avatar.show_text(cx, None, None, fallback_text);
-    }
-}
-
-impl Widget for RoomInfoSlidingPane {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        self.view.handle_event(cx, event, scope);
-
-        if !self.visible { return; }
-
-        let animator_action = self.animator_handle_event(cx, event);
-        if animator_action.must_redraw() {
-            self.redraw(cx);
-        }
-
-        if self.is_animating_out && !self.animator.is_track_animating(id!(panel)) {
-            self.visible = false;
-            self.is_animating_out = false;
-            cx.revert_key_focus();
-            self.view(cx, ids!(bg_view)).set_visible(cx, false);
-            self.redraw(cx);
-            return;
-        }
-
-        // Tap on the favourite star (in the hero) toggles the room's favourite
-        // tag. The star is a plain View (not a Button), so hit-test its area
-        // directly — this works in both inline and overlay modes.
-        let favorite_area = self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.name_row.favorite_button)).area();
-        if let Hit::FingerUp(fe) = event.hits(cx, favorite_area)
-            && fe.is_over && fe.is_primary_hit() && fe.was_tap()
-        {
-            // Flip the cached value first (ends the `self.info` borrow), then
-            // record the optimistic override + fire the async tag write.
-            let toggled = self.info.as_mut().map(|info| {
-                let new_is_favorite = !info.is_favorite;
-                info.is_favorite = new_is_favorite;
-                (info.owned_room_id.clone(), new_is_favorite)
-            });
-            if let Some((owned_room_id, new_is_favorite)) = toggled {
-                self.pending_favorite = Some((owned_room_id.clone(), new_is_favorite));
-                submit_async_request(MatrixRequest::SetIsFavorite {
-                    room_id: owned_room_id,
-                    is_favorite: new_is_favorite,
-                });
-                self.redraw(cx);
-            }
-        }
-
-        // Tap anywhere on the members card header opens the People sub-page.
-        let members_header_area = self.view(cx, ids!(content_scroll.info_view.members_card.members_header)).area();
-        if let Hit::FingerUp(fe) = event.hits(cx, members_header_area)
-            && fe.is_over && fe.is_primary_hit() && fe.was_tap()
-        {
-            self.show_people_page = true;
-            self.people_display_count = self.info.as_ref()
-                .map(|info| info.people_entries.len().min(40))
-                .unwrap_or(0);
-            // Always open the People list scrolled to the top (the PortalList
-            // otherwise keeps its previous scroll offset).
-            self.portal_list(cx, ids!(people_view.members_list_card.people_list))
-                .set_first_id_and_scroll(0, 0.0);
-            cx.widget_action(
-                self.widget_uid(),
-                RoomInfoPaneAction::ShowPeoplePage,
-            );
-            self.redraw(cx);
-        }
-
-        // Inline (tab) mode is opened/closed by the parent tab, so it never
-        // self-closes on back-press / tap-outside / Escape.
-        if !self.inline {
-            let area = self.view.area();
-            let close_pane = if is_invite_modal_open() || is_room_info_action_modal_open() {
-                matches!(
-                    event,
-                    Event::Actions(actions) if self.button(cx, ids!(close_button)).clicked(actions)
-                )
-            } else {
-                matches!(
-                    event,
-                    Event::Actions(actions) if self.button(cx, ids!(close_button)).clicked(actions)
-                )
-                || event.back_pressed()
-                || match event.hits_with_capture_overload(cx, area, true) {
-                    Hit::KeyUp(key) => key.key_code == KeyCode::Escape,
-                    Hit::FingerDown(_fde) => {
-                        cx.set_key_focus(area);
-                        false
-                    }
-                    Hit::FingerUp(fue) if fue.is_over => {
-                        fue.mouse_button().is_some_and(|b| b.is_back())
-                        || !self.view(cx, ids!(main_content)).area().rect(cx).contains(fue.abs)
-                    }
-                    _ => false,
-                }
-            };
-            if close_pane {
-                self.hide(cx);
-            }
-        }
-
-        if let Event::Actions(actions) = event {
-            // Re-bubble a person tap from one of OUR people rows as
-            // `OpenPeopleProfile` tagged with this pane's own widget uid, so the
-            // `RoomScreen` handler (which filters by pane uid) can pick it up.
-            //
-            // Gate on `show_people_page` so ONLY the instance currently showing
-            // its People page bubbles. Both the desktop overlay pane and the
-            // inline (mobile) pane receive every broadcast `Event::Actions`, and
-            // `PersonClicked` is deliberately a distinct variant that no pane
-            // ever emits — together this guarantees the two instances can never
-            // ping-pong the action between themselves (which froze the app).
-            if self.show_people_page {
-                for action in actions {
-                    if let RoomInfoPaneAction::PersonClicked(user_id) = action.as_widget_action().cast() {
-                        cx.widget_action(
-                            self.widget_uid(),
-                            RoomInfoPaneAction::OpenPeopleProfile(user_id.clone()),
-                        );
-                        break;
-                    }
-                }
-            }
-
-            if self.button(cx, ids!(header.back_button)).clicked(actions) {
-                self.show_people_page = false;
-                self.redraw(cx);
-            }
-            if self.button(cx, ids!(content_scroll.info_view.about_card.about_col.topic_toggle_button)).clicked(actions) {
-                self.topic_expanded = !self.topic_expanded;
-                self.redraw(cx);
-            }
-            if self.button(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.room_id_row.copy_room_id_button)).clicked(actions)
-                && let Some(info) = self.info.as_ref()
-            {
-                cx.copy_to_clipboard(&info.room_id);
-                enqueue_popup_notification(
-                    "Room ID copied.",
-                    PopupKind::Success,
-                    Some(2.0),
-                );
-            }
-            if self.button(cx, ids!(content_scroll.info_view.actions_row.actions_card.invite_button)).clicked(actions) {
-                cx.widget_action(
-                    self.widget_uid(),
-                    RoomInfoPaneAction::InviteUser,
-                );
-            }
-            if self.button(cx, ids!(content_scroll.info_view.actions_row.actions_card.report_room_button)).clicked(actions) {
-                cx.widget_action(
-                    self.widget_uid(),
-                    RoomInfoPaneAction::ReportRoom,
-                );
-            }
-            if self.button(cx, ids!(content_scroll.info_view.actions_row.leave_card.leave_room_button)).clicked(actions) {
-                cx.widget_action(
-                    self.widget_uid(),
-                    RoomInfoPaneAction::LeaveRoom,
-                );
-            }
-
-            if self.show_people_page
-                && let Some(info) = self.info.as_ref()
-                && self.people_display_count < info.people_entries.len()
-            {
-                let people_list = self.portal_list(cx, ids!(people_view.members_list_card.people_list));
-                if people_list.scrolled(actions) {
-                    let threshold = self.people_display_count.saturating_sub(5);
-                    if people_list.first_id() + people_list.visible_items() >= threshold {
-                        self.people_display_count = (self.people_display_count + 40).min(info.people_entries.len());
-                        self.redraw(cx);
-                    }
-                }
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let Some(info) = self.info.as_ref() else {
-            // In inline (tab) mode the parent controls visibility, so don't
-            // force-hide just because info hasn't been populated yet.
-            if !self.inline {
-                self.visible = false;
-            }
-            return self.view.draw_walk(cx, scope, walk);
-        };
-
-        // Slide animation + dimmed backdrop only apply to the overlay variant.
-        if !self.inline {
-            let panel_width = 320.0;
-            let right_margin = -(self.slide * panel_width);
-            let mut main_content = self.view(cx, ids!(main_content));
-            script_apply_eval!(cx, main_content, {
-                margin.right: #(right_margin)
-            });
-            let bg_alpha = (1.0 - self.slide) * 0.733;
-            let bg_color = vec4(0.0, 0.0, 0.0, bg_alpha);
-            let mut bg_view = self.view(cx, ids!(bg_view));
-            script_apply_eval!(cx, bg_view, {
-                draw_bg +: { color: #(bg_color) }
-            });
-        }
-
-        self.button(cx, ids!(header.back_button)).set_visible(cx, self.show_people_page);
-        // On the People sub-page the header shows the member count inline next to
-        // the back arrow; the plain "Info" title is blanked so they don't double up.
-        self.label(cx, ids!(header.title)).set_text(cx, if self.show_people_page { "" } else { "Info" });
-        self.label(cx, ids!(header.members_header_count)).set_visible(cx, self.show_people_page);
-        self.view(cx, ids!(content_scroll)).set_visible(cx, !self.show_people_page);
-        self.view(cx, ids!(content_scroll.info_view)).set_visible(cx, !self.show_people_page);
-        self.view(cx, ids!(people_view)).set_visible(cx, self.show_people_page);
-
-        // ----- Hero: name, room id, favourite star -----
-        self.label(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.name_row.title_wrap.room_name_value)).set_text(cx, &info.room_name);
-        // Bot pill trailing the room name, mirroring the rooms-list bot pill.
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.name_row.title_wrap.title_bot_pill)).set_visible(cx, info.show_title_bot_pill);
-        self.label(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.room_id_row.room_id_value)).set_text(cx, &info.room_id);
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.name_row.favorite_button.star_outline)).set_visible(cx, !info.is_favorite);
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.name_row.favorite_button.star_filled)).set_visible(cx, info.is_favorite);
-
-        // ----- Meta row: visibility / members / encryption -----
-        self.label(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.meta_row.visibility_meta.visibility_value)).set_text(cx, &info.visibility);
-        self.label(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.meta_row.members_meta.members_meta_value)).set_text(cx, &format!("{} members", info.member_count));
-        self.label(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.meta_row.encryption_meta.encryption_value)).set_text(cx, &info.encryption);
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.meta_row.encryption_meta.enc_icon_locked)).set_visible(cx, info.is_encrypted);
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.meta_row.encryption_meta.enc_icon_unlocked)).set_visible(cx, !info.is_encrypted);
-
-        // ----- Agent-enabled badge -----
-        self.view(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_meta.badges_row.agent_badge_wrap)).set_visible(cx, info.is_agent_enabled);
-
-        // ----- About / topic -----
-        let topic_chars_len = info.topic.chars().count();
-        let topic_has_more = topic_chars_len > TOPIC_PREVIEW_CHARS;
-        let topic_display_text = if topic_has_more && !self.topic_expanded {
-            let mut preview: String = info.topic.chars().take(TOPIC_PREVIEW_CHARS).collect();
-            preview.push_str("...");
-            preview
-        } else {
-            info.topic.clone()
-        };
-        self.label(cx, ids!(content_scroll.info_view.about_card.about_col.topic_value)).set_text(cx, &topic_display_text);
-        self.button(cx, ids!(content_scroll.info_view.about_card.about_col.topic_toggle_button)).set_visible(cx, topic_has_more);
-        self.button(cx, ids!(content_scroll.info_view.about_card.about_col.topic_toggle_button)).set_text(
-            cx,
-            if self.topic_expanded { "Collapse" } else { "Expand" },
-        );
-
-        // ----- Room hero avatar -----
-        let room_avatar = self.avatar(cx, ids!(content_scroll.info_view.summary_card.hero_row.room_avatar));
-        show_avatar_or_text(cx, &room_avatar, info.room_avatar_uri.as_ref(), &info.room_avatar_fallback_text);
-
-        // ----- Members card: count, avatar stack, your role -----
-        self.label(cx, ids!(content_scroll.info_view.members_card.members_header.members_count_group.members_count_value)).set_text(cx, &format!("{}", info.member_count));
-
-        let stack_shown = info.people_entries.len().min(3);
-        self.view(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_0)).set_visible(cx, stack_shown > 0);
-        self.view(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_1)).set_visible(cx, stack_shown > 1);
-        self.view(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_2)).set_visible(cx, stack_shown > 2);
-        if let Some(entry) = info.people_entries.first() {
-            let avatar = self.avatar(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_0.ring_0.stack_avatar_0));
-            show_avatar_or_text(cx, &avatar, entry.avatar_uri.as_ref(), &entry.avatar_fallback_text);
-        }
-        if let Some(entry) = info.people_entries.get(1) {
-            let avatar = self.avatar(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_1.ring_1.stack_avatar_1));
-            show_avatar_or_text(cx, &avatar, entry.avatar_uri.as_ref(), &entry.avatar_fallback_text);
-        }
-        if let Some(entry) = info.people_entries.get(2) {
-            let avatar = self.avatar(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_slot_2.ring_2.stack_avatar_2));
-            show_avatar_or_text(cx, &avatar, entry.avatar_uri.as_ref(), &entry.avatar_fallback_text);
-        }
-        let stack_more = info.people_entries.len().saturating_sub(stack_shown);
-        self.view(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_more_wrap)).set_visible(cx, stack_more > 0);
-        self.label(cx, ids!(content_scroll.info_view.members_card.members_detail.members_stack.stack_more_wrap.stack_more_chip.stack_more)).set_text(cx, &format!("+{stack_more}"));
-
-        self.view(cx, ids!(content_scroll.info_view.members_card.members_detail.my_role_wrap)).set_visible(cx, !info.my_role.is_empty());
-        self.label(cx, ids!(content_scroll.info_view.members_card.members_detail.my_role_wrap.my_role_chip.my_role_label)).set_text(cx, &info.my_role);
-
-        if self.show_people_page && self.people_display_count == 0 {
-            self.people_display_count = info.people_entries.len().min(40);
-        }
-        let visible_people_count = self.people_display_count.min(info.people_entries.len());
-        self.label(cx, ids!(header.members_header_count)).set_text(cx, &info.people_count_text);
-        self.view(cx, ids!(people_view.loading_label)).set_visible(cx, info.show_people_loading);
-        self.view(cx, ids!(people_view.empty_label)).set_visible(cx, !info.show_people_loading && info.people_entries.is_empty());
-        self.view(cx, ids!(people_view.members_list_card)).set_visible(cx, visible_people_count > 0);
-
-        while let Some(widget) = self.view.draw_walk(cx, scope, walk).step() {
-            let portal_list_ref = widget.as_portal_list();
-            let Some(mut list) = portal_list_ref.borrow_mut() else { continue };
-
-            list.set_item_range(cx, 0, visible_people_count);
-            while let Some(item_id) = list.next_visible_item(cx) {
-                let Some(entry) = info.people_entries.get(item_id) else { continue };
-                let item = list.item(cx, item_id, id!(PersonEntry));
-                item.as_room_info_people_entry().set_entry(cx, entry);
-                item.draw_all(cx, &mut Scope::empty());
-            }
-        }
-        DrawStep::done()
-    }
-}
-
-impl RoomInfoSlidingPane {
-    pub fn is_currently_shown(&self, _cx: &mut Cx) -> bool {
-        self.visible
-    }
-
-    fn set_info(&mut self, cx: &mut Cx, mut info: RoomInfoPaneInfo) {
-        // Switching to a DIFFERENT room must reset the sub-page state — otherwise
-        // a reused RoomScreen/pane lands on the previous room's People list
-        // (which is empty until the new room's members load). Always fall back to
-        // the Info view (and clear the topic-expand / paging state) on room change.
-        let room_changed = self.info.as_ref()
-            .is_some_and(|current| current.owned_room_id != info.owned_room_id);
-        if room_changed {
-            self.show_people_page = false;
-            self.people_display_count = 0;
-            self.topic_expanded = false;
-            self.portal_list(cx, ids!(people_view.members_list_card.people_list))
-                .set_first_id_and_scroll(0, 0.0);
-        }
-
-        // Preserve an in-flight optimistic favourite toggle across the frequent
-        // Signal-driven rebuilds: keep the optimistic value until the freshly
-        // built `info` (which re-reads `room.is_favourite()`) catches up, then
-        // drop the override. A different room invalidates any stale override.
-        if let Some((pending_room, pending_val)) = self.pending_favorite.clone() {
-            if pending_room == info.owned_room_id {
-                if info.is_favorite == pending_val {
-                    self.pending_favorite = None;
-                } else {
-                    info.is_favorite = pending_val;
-                }
-            } else {
-                self.pending_favorite = None;
-            }
-        }
-        self.info = Some(info);
-        if self.show_people_page {
-            if let Some(info) = self.info.as_ref() {
-                self.people_display_count = self.people_display_count
-                    .max(40.min(info.people_entries.len()))
-                    .min(info.people_entries.len());
-            }
-        }
-        self.redraw(cx);
-    }
-
-    pub fn show(&mut self, cx: &mut Cx) {
-        self.visible = true;
-        self.is_animating_out = false;
-        self.show_people_page = false;
-        self.topic_expanded = false;
-        self.people_display_count = 0;
-        cx.set_key_focus(self.view.area());
-        self.animator_play(cx, ids!(panel.show));
-        self.view(cx, ids!(bg_view)).set_visible(cx, true);
-        self.view.button(cx, ids!(close_button)).reset_hover(cx);
-        self.redraw(cx);
-    }
-
-    pub fn hide(&mut self, cx: &mut Cx) {
-        if !self.visible {
-            return;
-        }
-        self.is_animating_out = true;
-        self.animator_play(cx, ids!(panel.hide));
-        self.redraw(cx);
-    }
-}
-
-impl RoomInfoSlidingPaneRef {
-    pub fn is_currently_shown(&self, cx: &mut Cx) -> bool {
-        let Some(inner) = self.borrow() else { return false };
-        inner.is_currently_shown(cx)
-    }
-
-    fn set_info(&self, cx: &mut Cx, info: RoomInfoPaneInfo) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.set_info(cx, info);
-    }
-
-    /// Force the inline (docked, non-sliding) presentation. Set imperatively
-    /// because the DSL `inline: true` instance override is not reliably applied
-    /// to this `#[live]` field on the inline `info_content` instance.
-    pub fn set_inline(&self, inline: bool) {
-        if let Some(mut inner) = self.borrow_mut() {
-            inner.inline = inline;
-        }
-    }
-
-    pub fn show(&self, cx: &mut Cx) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.show(cx);
-    }
-
-    pub fn hide(&self, cx: &mut Cx) {
-        let Some(mut inner) = self.borrow_mut() else { return };
-        inner.hide(cx);
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum ReportRoomModalAction {
-    /// Emitted by RoomScreen to open the (now global, app-root) report modal
-    /// for a specific room. Carries the room so app.rs can route the result.
-    Open {
-        room_id: OwnedRoomId,
-        room_name_id: RoomNameId,
-    },
-    Close,
-    Submit(String),
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct ReportRoomModal {
-    #[deref]
-    view: View,
-    #[rust]
-    is_showing_error: bool,
-}
-
-impl Widget for ReportRoomModal {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        self.view.handle_event(cx, event, scope);
-        self.widget_match_event(cx, event, scope);
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-impl WidgetMatchEvent for ReportRoomModal {
-    fn handle_actions(&mut self, cx: &mut Cx, actions: &Actions, _scope: &mut Scope) {
-        let cancel_button = self.view.button(cx, ids!(buttons.cancel_button));
-        let report_button = self.view.button(cx, ids!(buttons.report_button));
-        let reason_input = self.view.text_input(cx, ids!(reason_input));
-        let mut status_label = self.view.label(cx, ids!(status_label));
-
-        if cancel_button.clicked(actions)
-            || actions
-                .iter()
-                .any(|a| matches!(a.downcast_ref(), Some(ModalAction::Dismissed)))
-        {
-            cx.action(ReportRoomModalAction::Close);
-            return;
-        }
-
-        if self.is_showing_error && reason_input.changed(actions).is_some() {
-            self.is_showing_error = false;
-            status_label.set_text(cx, "");
-            self.view.redraw(cx);
-        }
-
-        if report_button.clicked(actions) || reason_input.returned(actions).is_some() {
-            let reason = reason_input.text().trim().to_string();
-            if reason.is_empty() {
-                self.is_showing_error = true;
-                script_apply_eval!(cx, status_label, {
-                    text: "Please enter a reason before reporting."
-                    draw_text +: {
-                        color: mod.widgets.COLOR_FG_DANGER_RED
-                    }
-                });
-                self.view.redraw(cx);
-                return;
-            }
-            cx.action(ReportRoomModalAction::Submit(reason));
-        }
-    }
-}
-
-impl ReportRoomModal {
-    pub fn show(&mut self, cx: &mut Cx, room_name_id: &RoomNameId) {
-        self.is_showing_error = false;
-        self.view
-            .label(cx, ids!(title))
-            .set_text(cx, "Report Room");
-        self.view.label(cx, ids!(body)).set_text(
-            cx,
-            &format!(
-                "Report {} to your homeserver administrators. Please provide a reason.",
-                room_name_id
-            ),
-        );
-        self.view
-            .text_input(cx, ids!(reason_input))
-            .set_text(cx, "");
-        self.view.label(cx, ids!(status_label)).set_text(cx, "");
-        self.view
-            .button(cx, ids!(buttons.report_button))
-            .set_enabled(cx, true);
-        self.view
-            .button(cx, ids!(buttons.cancel_button))
-            .set_enabled(cx, true);
-        self.view
-            .button(cx, ids!(buttons.report_button))
-            .reset_hover(cx);
-        self.view
-            .button(cx, ids!(buttons.cancel_button))
-            .reset_hover(cx);
-        self.view.redraw(cx);
-    }
-}
-
-impl ReportRoomModalRef {
-    pub fn show(&self, cx: &mut Cx, room_name_id: &RoomNameId) {
-        let Some(mut inner) = self.borrow_mut() else {
-            return;
-        };
-        inner.show(cx, room_name_id);
-    }
+/// Registers this module's DSL blocks in dependency order: the sliding
+/// panes, the report modal, and the app-service panel first — the
+/// `RoomScreen` template in `dsl` references them via `mod.widgets.*`, and a
+/// template must already be registered when a later block resolves such a
+/// reference.
+pub fn script_mod(vm: &mut ScriptVm) {
+    threads_pane::script_mod(vm);
+    room_info_pane::script_mod(vm);
+    report_modal::script_mod(vm);
+    bot_admin::script_mod(vm);
+    dsl::script_mod(vm);
 }
 
 /// The main widget that displays a single Matrix room.
@@ -11846,186 +7484,6 @@ impl ActionDefaultRef for MessageAction {
     }
 }
 
-#[derive(Clone, Default, Debug)]
-pub enum AppServicePanelAction {
-    Dismiss,
-    OpenCreateBotModal,
-    OpenDeleteBotModal,
-    SendListBots,
-    SendBotHelp,
-    ShowBoundBots,
-    Unbind,
-    #[default]
-    None,
-}
-
-impl ActionDefaultRef for AppServicePanelAction {
-    fn default_ref() -> &'static Self {
-        static DEFAULT: AppServicePanelAction = AppServicePanelAction::None;
-        &DEFAULT
-    }
-}
-
-#[derive(Script, ScriptHook, Widget)]
-pub struct AppServicePanel {
-    #[deref] view: View,
-    #[rust] app_language: AppLanguage,
-    #[rust] app_language_initialized: bool,
-}
-
-impl Widget for AppServicePanel {
-    fn handle_event(&mut self, cx: &mut Cx, event: &Event, scope: &mut Scope) {
-        let app_language = scope.data.get::<AppState>()
-            .map(|app_state| app_state.app_language)
-            .unwrap_or_default();
-        if !self.app_language_initialized || self.app_language != app_language {
-            self.set_app_language(cx, app_language);
-        }
-        self.view.handle_event(cx, event, scope);
-
-        let room_screen_props = scope
-            .props
-            .get::<RoomScreenProps>()
-            .expect("BUG: RoomScreenProps should be available in Scope::props for AppServicePanel");
-        self.view
-            .button(cx, ids!(keyboard.third_row.view_bound_button))
-            .set_visible(cx, room_screen_props.app_service_enabled);
-        self.view
-            .button(cx, ids!(keyboard.third_row.unbind_button))
-            .set_visible(cx, room_screen_props.app_service_room_bound);
-
-        if let Event::Actions(actions) = event {
-            if self
-                .view
-                .button(cx, ids!(bubble.header.dismiss_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::Dismiss,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.first_row.create_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::OpenCreateBotModal,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.first_row.list_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::SendListBots,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.second_row.delete_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::OpenDeleteBotModal,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.second_row.help_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::SendBotHelp,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.third_row.view_bound_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::ShowBoundBots,
-                );
-            }
-
-            if self
-                .view
-                .button(cx, ids!(keyboard.third_row.unbind_button))
-                .clicked(actions)
-            {
-                cx.widget_action(
-                    room_screen_props.room_screen_widget_uid,
-                    AppServicePanelAction::Unbind,
-                );
-            }
-        }
-    }
-
-    fn draw_walk(&mut self, cx: &mut Cx2d, scope: &mut Scope, walk: Walk) -> DrawStep {
-        let app_language = scope.data.get::<AppState>()
-            .map(|app_state| app_state.app_language)
-            .unwrap_or_default();
-        if !self.app_language_initialized || self.app_language != app_language {
-            self.set_app_language(cx, app_language);
-        }
-        self.view.draw_walk(cx, scope, walk)
-    }
-}
-
-impl AppServicePanel {
-    fn set_app_language(&mut self, cx: &mut Cx, app_language: AppLanguage) {
-        self.app_language = app_language;
-        self.app_language_initialized = true;
-        self.view
-            .label(cx, ids!(sender_row.sender_name))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.sender_name"));
-        self.view
-            .label(cx, ids!(sender_row.sender_tag))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.sender_tag"));
-        self.view
-            .label(cx, ids!(bubble.header.title))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.title"));
-        self.view
-            .label(cx, ids!(bubble.subtitle))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.subtitle"));
-        self.view
-            .label(cx, ids!(bubble.footer.timestamp))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.timestamp_now"));
-        self.view
-            .button(cx, ids!(keyboard.first_row.create_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.create_bot"));
-        self.view
-            .button(cx, ids!(keyboard.first_row.list_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.list_bots"));
-        self.view
-            .button(cx, ids!(keyboard.second_row.delete_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.delete_bot"));
-        self.view
-            .button(cx, ids!(keyboard.second_row.help_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.bot_help"));
-        self.view
-            .button(cx, ids!(keyboard.third_row.view_bound_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.bots"));
-        self.view
-            .button(cx, ids!(keyboard.third_row.unbind_button))
-            .set_text(cx, tr_key(self.app_language, "room_screen.app_service.button.unbind"));
-        self.view.redraw(cx);
-    }
-}
-
 /// A widget representing a single message of any kind within a room timeline.
 #[derive(Script, ScriptHook, Widget, Animator)]
 pub struct Message {
@@ -12434,226 +7892,12 @@ mod tests {
     }
 
     #[test]
-    fn test_room_info_bot_identity_fingerprint_tracks_registry_agents() {
-        let agent_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-        let before = room_info_bot_identity_fingerprint(Some(&app_state), None);
-
-        app_state
-            .agent_registry
-            .register(agent_id.clone(), crate::app::AgentEntry::default());
-        let after = room_info_bot_identity_fingerprint(Some(&app_state), None);
-
-        assert_ne!(before, after);
-        assert!(after.known_bot_user_ids.iter().any(|id| id == &agent_id));
-    }
-
-    #[test]
-    fn test_room_info_bot_identity_fingerprint_tracks_appservice_known_bots() {
-        let current_user_id: OwnedUserId = "@alice:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-        app_state.bot_settings.enabled = true;
-        let bot_id = app_state
-            .bot_settings
-            .resolved_bot_user_id(Some(current_user_id.as_ref()))
-            .unwrap();
-        let before = room_info_bot_identity_fingerprint(
-            Some(&app_state),
-            Some(current_user_id.as_ref()),
-        );
-
-        app_state.bot_settings.record_known_bot_user_ids([bot_id.clone()]);
-        let after = room_info_bot_identity_fingerprint(
-            Some(&app_state),
-            Some(current_user_id.as_ref()),
-        );
-
-        assert_ne!(before, after);
-        assert_eq!(after.resolved_parent_bot_user_id.as_ref(), Some(&bot_id));
-        assert!(after.known_bot_user_ids.iter().any(|id| id == &bot_id));
-    }
-
-    #[test]
-    fn test_room_info_bot_marker_hidden_after_agentlab_unbind() {
-        let current_user_id: OwnedUserId = "@alice:example.org".try_into().unwrap();
-        let agent_id: OwnedUserId = "@octos_mac:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-
-        app_state.agent_registry.register(agent_id.clone(), crate::app::AgentEntry {
-            framework: crate::app::AgentFramework::Octos,
-            ..Default::default()
-        });
-        app_state.bot_settings.enabled = true;
-        app_state.bot_settings.botfather_user_id = agent_id.to_string();
-        app_state.bot_settings.record_known_bot_user_ids([agent_id.clone()]);
-
-        let before = room_info_bot_identity_fingerprint(
-            Some(&app_state),
-            Some(current_user_id.as_ref()),
-        );
-        assert!(is_known_or_likely_bot(
-            agent_id.as_ref(),
-            before.resolved_parent_bot_user_id.as_deref(),
-            &before.known_bot_user_ids,
-        ));
-
-        app_state.unregister_agent_and_clear_bot_identity(
-            agent_id.as_ref(),
-            Some(current_user_id.as_ref()),
-        );
-        let after = room_info_bot_identity_fingerprint(
-            Some(&app_state),
-            Some(current_user_id.as_ref()),
-        );
-
-        assert!(after.resolved_parent_bot_user_id.is_none());
-        assert!(after.known_bot_user_ids.is_empty());
-        assert!(!is_known_or_likely_bot(
-            agent_id.as_ref(),
-            after.resolved_parent_bot_user_id.as_deref(),
-            &after.known_bot_user_ids,
-        ));
-    }
-
-    #[test]
     fn test_room_members_fetch_updates_rooms_list_member_ids() {
         let src = include_str!("mod.rs");
 
         assert!(src.contains("TimelineUpdate::RoomMembersListFetched"));
         assert!(src.contains("RoomsListUpdate::UpdateRoomMemberUserIds"));
         assert!(src.contains("member.user_id().to_owned()"));
-    }
-
-    #[test]
-    fn test_room_info_title_bot_pill_hidden_after_room_unbound() {
-        let room_id: OwnedRoomId = "!room:example.org".try_into().unwrap();
-        let bot_id: OwnedUserId = "@bot:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-
-        app_state
-            .bot_settings
-            .set_room_bound(room_id.clone(), Some(bot_id.clone()), true);
-        assert!(room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, std::iter::empty(),
-        ));
-
-        app_state
-            .bot_settings
-            .set_room_bound(room_id.clone(), Some(bot_id), false);
-        assert!(!room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, std::iter::empty(),
-        ));
-    }
-
-    #[test]
-    fn test_room_info_title_bot_pill_shown_when_member_is_registered_agent() {
-        let room_id: OwnedRoomId = "!group:example.org".try_into().unwrap();
-        let agent_id: OwnedUserId = "@octos_mac:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-
-        app_state
-            .agent_registry
-            .register(agent_id.clone(), crate::app::AgentEntry::default());
-        assert!(room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, [agent_id.as_ref()],
-        ));
-
-        app_state.agent_registry.unregister(agent_id.as_ref());
-        assert!(!room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, [agent_id.as_ref()],
-        ));
-    }
-
-    #[test]
-    fn test_room_info_title_bot_pill_hidden_after_agent_registry_unbind() {
-        let room_id: OwnedRoomId = "!dm:example.org".try_into().unwrap();
-        let agent_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-
-        app_state
-            .agent_registry
-            .register(agent_id.clone(), crate::app::AgentEntry::default());
-        assert!(room_info_title_shows_agent_badge(
-            Some(&app_state),
-            room_id.as_ref(),
-            Some(agent_id.as_ref()),
-            std::iter::empty(),
-        ));
-
-        app_state.agent_registry.unregister(agent_id.as_ref());
-        assert!(!room_info_title_shows_agent_badge(
-            Some(&app_state),
-            room_id.as_ref(),
-            Some(agent_id.as_ref()),
-            std::iter::empty(),
-        ));
-    }
-
-    #[test]
-    fn test_room_info_title_bot_pill_hidden_after_agentlab_unbind_clears_binding() {
-        let current_user_id: OwnedUserId = "@alice:example.org".try_into().unwrap();
-        let room_id: OwnedRoomId = "!room:example.org".try_into().unwrap();
-        let agent_id: OwnedUserId = "@octos_mac:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-
-        app_state.agent_registry.register(agent_id.clone(), crate::app::AgentEntry {
-            framework: crate::app::AgentFramework::Octos,
-            ..Default::default()
-        });
-        app_state.bot_settings.enabled = true;
-        app_state.bot_settings.botfather_user_id = agent_id.to_string();
-        app_state.bot_settings.record_known_bot_user_ids([agent_id.clone()]);
-        app_state
-            .bot_settings
-            .set_room_bound(room_id.clone(), Some(agent_id.clone()), true);
-        assert!(room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, std::iter::empty(),
-        ));
-
-        app_state.unregister_agent_and_clear_bot_identity(
-            agent_id.as_ref(),
-            Some(current_user_id.as_ref()),
-        );
-
-        assert!(!room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, std::iter::empty(),
-        ));
-    }
-
-    #[test]
-    fn test_room_info_title_bot_pill_ignores_known_bot_without_binding_or_agent_dm() {
-        let room_id: OwnedRoomId = "!group:example.org".try_into().unwrap();
-        let bot_id: OwnedUserId = "@bot:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-        app_state.bot_settings.record_known_bot_user_ids([bot_id.clone()]);
-
-        // A known bot that is neither bound, a DM target, nor a registered-agent
-        // member must not trigger the pill — even when it is a room member.
-        assert!(!room_info_title_shows_agent_badge(
-            Some(&app_state), room_id.as_ref(), None, [bot_id.as_ref()],
-        ));
-    }
-
-    #[test]
-    fn test_room_info_dm_target_requires_single_non_self_member() {
-        let me: OwnedUserId = "@me:example.org".try_into().unwrap();
-        let agent_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let human_id: OwnedUserId = "@human:example.org".try_into().unwrap();
-
-        assert_eq!(
-            room_info_dm_target_from_user_ids(
-                [me.as_ref(), agent_id.as_ref()],
-                Some(me.as_ref()),
-            ),
-            Some(agent_id.clone()),
-        );
-        assert_eq!(
-            room_info_dm_target_from_user_ids(
-                [me.as_ref(), agent_id.as_ref(), human_id.as_ref()],
-                Some(me.as_ref()),
-            ),
-            None,
-        );
     }
 
 }
